### 2020-07-10 #2468 @kyleniemeyer
Hi @zingale, thanks for your submission!

Before we get started, I just want to check that you haven't already published a software paper describing CASTRO. (I recall seeing some papers about CASTRO in the past.) Papers using CASTRO are fine, but if you have already published a software article about CASTRO, then we would consider this to be a duplication. Exceptions would be if this represents a substantially new/different version.
