### 2020-07-10 #2468 @zingale
I guess this depends on the difference between a code paper and an algorithm paper.  We've mostly published on new algorithm development and performance and optimization.  The only paper that covers all of Castro in general is:

* CASTRO: A New Compressible Astrophysical Solver. I. Hydrodynamics and Self-gravity, 2010, ApJ:

  https://ui.adsabs.harvard.edu/abs/2010ApJ...715.1221A/abstract

There was a paper II and III in that series that focused on radiation and neutrino transport.  The code has changed a lot since 2010, and many new developers have joined that do not receive credit from citing that 2010 paper.

We've had a number of conference proceedings and focused algorithm papers:

* MAESTRO, CASTRO, and SEDONA -- Petascale Codes for Astrophysical Applications, SciDAC 2010

  https://ui.adsabs.harvard.edu/abs/2010arXiv1008.2801A/abstract

  This is quite old and doesn't go into much detail

* On the Piecewise Parabolic Method for Compressible Flow With Stellar Equations of State, ApJ, 2015

  https://ui.adsabs.harvard.edu/abs/2015ApJS..216...31Z/abstract

  This is a pure algorithm paper that focuses only on the reconstruction in the hydro for a general EOS, so is very limited in
  scope.

* Meeting the Challenges of Modeling Astrophysical Thermonuclear Explosions: Castro, Maestro, and the AMReX Astrophysics Suite,  AstroNum 2017

  https://ui.adsabs.harvard.edu/abs/2018JPhCS1031a2024Z/abstract

  This mainly talks about applications and shows CPU / GPU scaling curves for Castro (for an early version of our GPU effort)

* Toward Resolved Simulations of Burning Fronts in Thermonuclear X-ray Bursts, AstroNum 2018

  https://iopscience.iop.org/article/10.1088/1742-6596/1225/1/012005

  This discusses Castro in the context of X-ray burst science

* The Castro AMR Simulation Code: Current and Future Developments, AstroNum 2019 (submitted)

  https://ui.adsabs.harvard.edu/abs/2019arXiv191012578Z/abstract

  This summarizes our new time-integration and GPU performance

* Improved Coupling of Hydrodynamics and Nuclear Reactions via Spectral Deferred Corrections

  https://ui.adsabs.harvard.edu/abs/2019ApJ...886..105Z/abstract

  This is an algorithm paper that describes our new time-integration strategy.

* Preparing Nuclear Astrophysics for Exascale, accepted to SC20 (should appear on astro-ph next week)

  This is our GPU paper that describes our GPU strategy.

In addition to those, there are a number of science papers that contain appendices that include some verification of a particular
part of the algorithm.

The reason we are submitting to JOSS is there is no up-to-date paper to cite for the present version of the code from the current developer groups.  In general, we only cite the 2010 ApJ paper in our science paper.

We've never had a paper where the reviewers looked at the source code, documentation, and verified that it ran as we describe, which is our motivation for submitting to JOSS.
