### 2020-07-11 c70dd7c
Calculate rotational acceleration inline and remove Rotation_Type
