+++ Source/rotation/Castro_rotation.cpp
@@ -13,20 +13,18 @@ Castro::construct_old_rotation_source(MultiFab& source, MultiFab& state_in, Real
     const Real strt_time = ParallelDescriptor::second();
 
     MultiFab& phirot_old = get_old_data(PhiRot_Type);
-    MultiFab& rot_old = get_old_data(Rotation_Type);
 
     // Fill the rotation data.
 
     if (!do_rotation) {
 
         phirot_old.setVal(0.0);
-        rot_old.setVal(0.0);
 
         return;
 
     }
 
-    fill_rotation_field(phirot_old, rot_old, state_in, time);
+    fill_rotation_field(phirot_old, state_in, time);
 
     const Real *dx = geom.CellSize();
     const int* domlo = geom.Domain().loVect();
@@ -39,7 +37,7 @@ Castro::construct_old_rotation_source(MultiFab& source, MultiFab& state_in, Real
     {
         const Box& bx = mfi.tilebox();
 
-        rsrc(bx, rot_old.array(mfi), state_in.array(mfi), source.array(mfi), dt);
+        rsrc(bx, state_in.array(mfi), source.array(mfi), dt);
 
     }
 
@@ -71,23 +69,20 @@ Castro::construct_new_rotation_source(MultiFab& source, MultiFab& state_old, Mul
     const Real strt_time = ParallelDescriptor::second();
 
     MultiFab& phirot_old = get_old_data(PhiRot_Type);
-    MultiFab& rot_old = get_old_data(Rotation_Type);
 
     MultiFab& phirot_new = get_new_data(PhiRot_Type);
-    MultiFab& rot_new = get_new_data(Rotation_Type);
 
     // Fill the rotation data.
 
     if (!do_rotation) {
 
         phirot_new.setVal(0.);
-        rot_new.setVal(0.);
 
         return;
 
     }
 
-    fill_rotation_field(phirot_new, rot_new, state_new, time);
+    fill_rotation_field(phirot_new, state_new, time);
 
     // Now do corrector part of rotation source term update
 
@@ -105,7 +100,6 @@ Castro::construct_new_rotation_source(MultiFab& source, MultiFab& state_old, Mul
 
             corrrsrc(bx,
                      phirot_old.array(mfi), phirot_new.array(mfi),
-                     rot_old.array(mfi), rot_new.array(mfi),
                      state_old.array(mfi), state_new.array(mfi),
                      source.array(mfi),
                      (*mass_fluxes[0]).array(mfi), (*mass_fluxes[1]).array(mfi), (*mass_fluxes[2]).array(mfi),
@@ -133,7 +127,7 @@ Castro::construct_new_rotation_source(MultiFab& source, MultiFab& state_old, Mul
 
 
 
-void Castro::fill_rotation_field(MultiFab& phi, MultiFab& rot, MultiFab& state_in, Real time)
+void Castro::fill_rotation_field(MultiFab& phi, MultiFab& state_in, Real time)
 {
 
     BL_PROFILE("Castro::fill_rotation_field()");
@@ -153,25 +147,6 @@ void Castro::fill_rotation_field(MultiFab& phi, MultiFab& rot, MultiFab& state_i
         const Box& bx = mfi.growntilebox(ng);
 
         fill_rotational_potential(bx, phi.array(mfi), time);
-
-    }
-
-    rot.setVal(0.0);
-
-    ng = state_in.nGrow();
-
-    if (ng > rot.nGrow())
-        amrex::Error("State MF has more ghost cells than rotation MF.");
-
-#ifdef _OPENMP
-#pragma omp parallel
-#endif
[…]
