### 2020-07-11 a53843a
Convert wdmerger derives to C++
