+++ Exec/science/wdmerger/Problem_Derive.H
@@ -1,153 +1,80 @@
 
-#ifdef __cplusplus
-extern "C"
-{
-#endif
+void ca_derinertialmomentumx
+(const amrex::Box& bx, amrex::FArrayBox& derfab, int dcomp, int /*ncomp*/,
+ const amrex::FArrayBox& datfab, const amrex::Geometry& geomdata,
+ amrex::Real /*time*/, const int* /*bcrec*/, int /*level*/);
 
-  void ca_derinertialmomentumx
-    (BL_FORT_FAB_ARG_3D(der), const int* nvar,
-     const BL_FORT_FAB_ARG_3D(data), const int* ncomp,
-     const int* lo, const int* hi,
-     const int* domain_lo, const int* domain_hi,
-     const amrex::Real* delta, const amrex::Real* xlo,
-     const amrex::Real* time, const amrex::Real* dt, const int* bcrec, 
-     const int* level, const int* grid_no);
+void ca_derinertialmomentumy
+(const amrex::Box& bx, amrex::FArrayBox& derfab, int dcomp, int /*ncomp*/,
+ const amrex::FArrayBox& datfab, const amrex::Geometry& geomdata,
+ amrex::Real /*time*/, const int* /*bcrec*/, int /*level*/);
 
-  void ca_derinertialmomentumy
-    (BL_FORT_FAB_ARG_3D(der), const int* nvar,
-     const BL_FORT_FAB_ARG_3D(data), const int* ncomp,
-     const int* lo, const int* hi,
-     const int* domain_lo, const int* domain_hi,
-     const amrex::Real* delta, const amrex::Real* xlo,
-     const amrex::Real* time, const amrex::Real* dt, const int* bcrec, 
-     const int* level, const int* grid_no);
+void ca_derinertialmomentumz
+(const amrex::Box& bx, amrex::FArrayBox& derfab, int dcomp, int /*ncomp*/,
+ const amrex::FArrayBox& datfab, const amrex::Geometry& geomdata,
+ amrex::Real /*time*/, const int* /*bcrec*/, int /*level*/);
 
-  void ca_derinertialmomentumz
-    (BL_FORT_FAB_ARG_3D(der), const int* nvar,
-     const BL_FORT_FAB_ARG_3D(data), const int* ncomp,
-     const int* lo, const int* hi,
-     const int* domain_lo, const int* domain_hi,
-     const amrex::Real* delta, const amrex::Real* xlo,
-     const amrex::Real* time, const amrex::Real* dt, const int* bcrec, 
-     const int* level, const int* grid_no);
+void ca_derinertialangmomx
+(const amrex::Box& bx, amrex::FArrayBox& derfab, int dcomp, int /*ncomp*/,
+ const amrex::FArrayBox& datfab, const amrex::Geometry& geomdata,
+ amrex::Real /*time*/, const int* /*bcrec*/, int /*level*/);
 
-  void ca_derinertialangmomx
-    (BL_FORT_FAB_ARG_3D(der), const int* nvar,
-     const BL_FORT_FAB_ARG_3D(data), const int* ncomp,
-     const int* lo, const int* hi,
-     const int* domain_lo, const int* domain_hi,
-     const amrex::Real* delta, const amrex::Real* xlo,
-     const amrex::Real* time, const amrex::Real* dt, const int* bcrec, 
-     const int* level, const int* grid_no);
+void ca_derinertialangmomy
+(const amrex::Box& bx, amrex::FArrayBox& derfab, int dcomp, int /*ncomp*/,
+ const amrex::FArrayBox& datfab, const amrex::Geometry& geomdata,
+ amrex::Real /*time*/, const int* /*bcrec*/, int /*level*/);
 
[…]

[…]
