### 2020-07-11 #2468 @arfon
@zingale - does this paper accompany a submission to a AAS journal, i.e. is this part of https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
