### 2020-07-11 58e4ff0
Remove redundant variables
