### 2020-07-11 86827fa
Fix primary and secondary masks
