### 2020-07-11 7bbc9ec
Don't subtract off center
