### 2020-07-11 d06a6e4
Fix reduction ordering
