### 2020-07-11 c27fe90
Convert wdmerger derives to C++ (#1129)
