+++ Exec/science/wdmerger/Problem_Derive.cpp
@@ -1,312 +1,681 @@
 #include "AMReX_REAL.H"
 
 #include "Derive.H"
-#include "Problem_Derive_F.H"
 #include "Castro.H"
 #include "Castro_F.H"
+#include <fundamental_constants.H>
+#include <prob_parameters.H>
+#include <wdmerger_util.H>
+#include <wdmerger_data.H>
+#ifdef ROTATION
+#include "Castro_rotation_F.H"
+#endif
 
 using namespace amrex;
 
-#ifdef __cplusplus
-extern "C"
+void ca_derinertialmomentumx(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
+                             const FArrayBox& datfab, const Geometry& geomdata,
+                             Real /*time*/, const int* /*bcrec*/, int /*level*/)
 {
+    // Derive momentum, given the grid momenta.
+
+    auto const dat = datfab.array();
+    auto const der = derfab.array();
+
+    const auto dx = geomdata.CellSizeArray();
+    const auto problo = geomdata.ProbLoArray();
+
+    GpuArray<Real, 3> center;
+    ca_get_center(center.begin());
+
+    GpuArray<Real, 3> omega;
+    get_omega(omega.begin());
+
+    amrex::ParallelFor(bx,
+    [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
+    {
+        GpuArray<Real, 3> loc;
+        loc[0] = problo[0] + (static_cast<Real>(i) + 0.5_rt) * dx[0] - center[0];
+
+#if AMREX_SPACEDIM >= 2
+        loc[1] = problo[1] + (static_cast<Real>(j) + 0.5_rt) * dx[1] - center[1];
+#else
+        loc[1] = 0.0_rt;
 #endif
 
-    // Note that in the following routines, we are NOT passing
-    // several variables to Fortran that would be unused.
+#if AMREX_SPACEDIM == 3
+        loc[2] = problo[2] + (static_cast<Real>(k) + 0.5_rt) * dx[2] - center[2];
+#else
+        loc[2] = 0.0_rt;
+#endif
 
-    // These routines are called in an MFIter loop, so we do not
-    // need to explicitly synchronize after GPU kernels.
-
-    void ca_derinertialmomentumx(Real* der, const int* der_lo, const int* der_hi, const int* nvar,
-                                 const Real* data, const int* data_lo, const int* data_hi, const int* ncomp,
-                                 const int* lo, const int* hi,
-                                 const int* domain_lo, const int* domain_hi,
-                                 const Real* dx, const Real* xlo,
-                                 const Real* time, const Real* dt, const int* bcrec, 
-                                 const int* level, const int* grid_no)
-    {
-
-#pragma gpu
-        derinertialmomentumx(der, AMREX_INT_ANYD(der_lo), AMREX_INT_ANYD(der_hi), *nvar,
-                             data, AMREX_INT_ANYD(data_lo), AMREX_INT_ANYD(data_hi), *ncomp,
-                             AMREX_INT_ANYD(lo), AMREX_INT_ANYD(hi),
-                             AMREX_INT_ANYD(domain_lo), AMREX_INT_ANYD(domain_hi),
-                             AMREX_REAL_ANYD(dx), *time);
-
-    }
-
-    void ca_derinertialmomentumy(Real* der, const int* der_lo, const int* der_hi, const int* nvar,
-                                 const Real* data, const int* data_lo, const int* data_hi, const int* ncomp,
-                                 const int* lo, const int* hi,
-                                 const int* domain_lo, const int* domain_hi,
-                                 const Real* dx, const Real* xlo,
-                                 const Real* time, const Real* dt, const int* bcrec, 
-                                 const int* level, const int* grid_no)
[…]
