### 2020-07-11 00e8962
Implement conservative rotations on edges like gravity
