+++ Source/rotation/rotation_sources.cpp
@@ -180,6 +180,16 @@ Castro::corrrsrc(const Box& bx,
   // Note that the time passed to this function
   // is the new time at time-level n+1.
 
+  GpuArray<Real, 3> dx;
+  for (int i = 0; i < AMREX_SPACEDIM; ++i) {
+      dx[i] = geom.CellSizeArray()[i];
+  }
+  for (int i = AMREX_SPACEDIM; i < 3; ++i) {
+      dx[i] = 0.0_rt;
+  }
+
+  Real hdtInv = 0.5_rt / dt;
+
   GpuArray<Real, 3> center;
   ca_get_center(center.begin());
 
@@ -456,41 +466,51 @@ Castro::corrrsrc(const Box& bx,
 
       SrEcorr = - SrE_old;
 
-      // The change in the gas energy is equal in magnitude to, and opposite in sign to,
-      // the change in the rotational potential energy, rho * phi.
-      // This must be true for the total energy, rho * E_gas + rho * phi, to be conserved.
-      // Consider as an example the zone interface i+1/2 in between zones i and i + 1.
-      // There is an amount of mass drho_{i+1/2} leaving the zone. From this zone's perspective
-      // it starts with a potential phi_i and leaves the zone with potential phi_{i+1/2} =
-      // (1/2) * (phi_{i-1}+phi_{i}). Therefore the new rotational energy is equal to the mass
-      // change multiplied by the difference between these two potentials.
-      // This is a generalization of the cell-centered approach implemented in
-      // the other source options, which effectively are equal to
-      // SrEcorr = - drho(i,j,k) * phi(i,j,k),
-      // where drho(i,j,k) = HALF * (unew(i,j,k,URHO) - uold(i,j,k,URHO)).
+      // See corrgsrc for an explanation of this algorithm. We can implement this identically
+      // to the conservative gravity, swapping out the gravitational acceleration for the
+      // rotational acceleration. We only need to calculate the rotational acceleration on
+      // edges, which is straightforward to do. The Coriolis term does not play a role in
+      // the energy update (it does no work), so we turn it off and pass in a temporary
+      // velocity array which will be ignored.
 
-      // Note that in the hydrodynamics step, the fluxes used here were already
-      // multiplied by dA and dt, so dividing by the cell volume is enough to
-      // get the density change (flux * dt * dA / dV). We then divide by dt
-      // so that we get the source term and not the actual update, which will
-      // be applied later by multiplying by dt.
+      Real edge_Sr[3][2];
 
-      Real phi = 0.5_rt * (phi_new(i,j,k) + phi_old(i,j,k));
+      // Loop over directions and edges
 
-      Real phixl = 0.5_rt * (phi_new(i-1,j,k) + phi_old(i-1,j,k));
-      Real phixr = 0.5_rt * (phi_new(i+1,j,k) + phi_old(i+1,j,k));
-      Real phiyl = 0.5_rt * (phi_new(i,j-dg1,k) + phi_old(i,j-dg1,k));
-      Real phiyr = 0.5_rt * (phi_new(i,j+dg1,k) + phi_old(i,j+dg1,k));
-      Real phizl = 0.5_rt * (phi_new(i,j,k-dg2) + phi_old(i,j,k-dg2));
-      Real phizr = 0.5_rt * (phi_new(i,j,k+dg2) + phi_old(i,j,k+dg2));
+      for (int dir = 0; dir < 3; ++dir) {
+          for (int edge = 0; edge <= 1; ++edge) {
 
-      SrEcorr = SrEcorr - (0.5_rt / dt) * ( flux0(i    ,j,k) * (phi - phixl) -
[…]
