### 2020-07-17 caf8a2a
Move hydro device functions to headers
