+++ Source/hydro/riemann_solvers.cpp
@@ -1299,148 +1299,3 @@ Castro::HLLC(const Box& bx,
   });
 }
 
-AMREX_GPU_HOST_DEVICE
-void
-Castro::HLL(const Real* ql, const Real* qr,
-            const Real cl, const Real cr,
-            const int idir, const int coord,
-            Real* flux_hll) {
-
-  // This is the HLLE solver.  We should apply it to zone averages
-  // (not reconstructed states) at an interface in the presence of
-  // shocks to avoid the odd-even decoupling / carbuncle phenomenon.
-  //
-  // See: Einfeldt, B.  et al. 1991, JCP, 92, 273
-  //      Einfeldt, B. 1988, SIAM J NA, 25, 294
-
-
-  constexpr Real small_hll = 1.e-10_rt;
-
-  int ivel, ivelt, iveltt;
-  int imom, imomt, imomtt;
-
-  if (idir == 0) {
-    ivel = QU;
-    ivelt = QV;
-    iveltt = QW;
-
-    imom = UMX;
-    imomt = UMY;
-    imomtt = UMZ;
-
-  } else if (idir == 1) {
-    ivel = QV;
-    ivelt = QU;
-    iveltt = QW;
-
-    imom = UMY;
-    imomt = UMX;
-    imomtt = UMZ;
-
-  } else {
-    ivel = QW;
-    ivelt = QU;
-    iveltt = QV;
-
-    imom = UMZ;
-    imomt = UMX;
-    imomtt = UMY;
-  }
-
-  Real rhol_sqrt = std::sqrt(ql[QRHO]);
-  Real rhor_sqrt = std::sqrt(qr[QRHO]);
-
-  Real rhod = 1.0_rt/(rhol_sqrt + rhor_sqrt);
-
-
-  // compute the average sound speed. This uses an approximation from
-  // E88, eq. 5.6, 5.7 that assumes gamma falls between 1
-  // and 5/3
-  Real cavg = std::sqrt( (rhol_sqrt*cl*cl + rhor_sqrt*cr*cr)*rhod +
-                         0.5_rt*rhol_sqrt*rhor_sqrt*rhod*rhod*std::pow(qr[ivel] - ql[ivel], 2));
-
-
-  // Roe eigenvalues (E91, eq. 5.3b)
-  Real uavg = (rhol_sqrt*ql[ivel] + rhor_sqrt*qr[ivel])*rhod;
-
-  Real a1 = uavg - cavg;
-  Real a4 = uavg + cavg;
-
-
-  // signal speeds (E91, eq. 4.5)
-  Real bl = amrex::min(a1, ql[ivel] - cl);
-  Real br = amrex::max(a4, qr[ivel] + cr);
-
-  Real bm = amrex::min(0.0_rt, bl);
-  Real bp = amrex::max(0.0_rt, br);
-
-  Real bd = bp - bm;
-
-  if (std::abs(bd) < small_hll*amrex::max(std::abs(bm), std::abs(bp))) return;
-
-  // we'll overwrite the passed in flux with the HLL flux
-
-  bd = 1.0_rt/bd;
-
-  // compute the fluxes according to E91, eq. 4.4b -- note that the
-  // min/max above picks the correct flux if we are not in the star
-  // region
-
-  // density flux
[…]
