### 2020-07-17 3d5911b
initialize aux
