### 2020-07-17 9bd57b1
fix header + renormalize species
