### 2020-07-17 bccac66
if we are doing NSE, then abar was getting added to the derives twice
