+++ Source/driver/Castro_setup.cpp
@@ -900,12 +900,15 @@ Castro::variableSetUp ()
     derive_lst.addComponent(spec_string,desc_lst,State_Type,UFS+i,1);
   }
 
+#ifndef NSE_THERMO
   //
   // Abar
+  // note: if we are using NSE thermodynamics, then abar is already an aux quantity
   //
   derive_lst.add("abar",IndexType::TheCellType(),1,ca_derabar,the_same_box);
   derive_lst.addComponent("abar",desc_lst,State_Type,URHO,1);
   derive_lst.addComponent("abar",desc_lst,State_Type,UFS,NumSpec);
+#endif
 
   //
   // Velocities
