### 2020-07-17 22ebe33
fix the indexing into rho_auxdot for the derived (#1138)
