+++ paper-parser/papers.bib
@@ -1142,18 +1142,21 @@ archivePrefix = "arXiv",
          {Jones}, David O. and {Kirshner}, Robert P. and {Pan}, Yen-Chen and
          {Piro}, Anthony L. and {Rest}, Armin and {Rojas-Bravo}, Cesar},
         title = "{Ca hnk: Calcium-rich Transient SN 2016hnk from the Helium Shell Detonation of a Sub-Chandrasekhar White Dwarf}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - High Energy Astrophysical Phenomena},
-         year = "2019",
-        month = "Oct",
-          eid = {arXiv:1910.05436},
-        pages = {arXiv:1910.05436},
+      journal = {\apj},
+     keywords = {UAT:1799, UAT:1728, UAT:1145, UAT:1668, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2020,
+        month = "jun",
+       volume = {896},
+       number = {2},
+          eid = {165},
+        pages = {165},
+          doi = {10.3847/1538-4357/ab94b8},
 archivePrefix = {arXiv},
        eprint = {1910.05436},
  primaryClass = {astro-ph.HE},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv191005436J},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...896..165J},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-  subject = {Type Ia supernovae&mdash;sub-Chandra model}
+      subject = {Type Ia supernovae&mdash;sub-Chandra model}
 }
 
 @ARTICLE{2020arXiv200501782J,
@@ -1237,3 +1240,22 @@ archivePrefix = {arXiv},
       subject = {X-ray bursts}
 }
 
+@ARTICLE{2020arXiv200705218K,
+       author = {{Katz}, Max P. and {Almgren}, Ann and {Barrios Sazo}, Maria and
+         {Eiden}, Kiran and {Gott}, Kevin and {Harpole}, Alice and
+         {Sexton}, Jean M. and {Willcox}, Don E. and {Zhang}, Weiqun and
+         {Zingale}, Michael},
+        title = "{Preparing Nuclear Astrophysics for Exascale}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2020,
+        month = "jul",
+          eid = {arXiv:2007.05218},
+        pages = {arXiv:2007.05218},
+archivePrefix = {arXiv},
+       eprint = {2007.05218},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200705218K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  subject = {algorithm}
+}
