### 2020-07-23 #2513 @whedon
[REVIEW]: CASTRO: A Massively Parallel Compressible Astrophysics Simulation Code
**Submitting author:** @zingale (<a href="http://orcid.org/0000-0001-8401-030X">M. Zingale</a>)
**Repository:** <a href="https://github.com/amrex-astro/Castro" target ="_blank">https://github.com/amrex-astro/Castro</a>
**Version:** JOSS_paper
