### 2020-07-23 635f564
address some compiler warnings
