### 2020-07-23 #2513 @zingale
that proof looks good
