### 2020-07-24 1e1004e
C++ doesn't like size 0 arrays for eos_t's aux block them out if NAUX_NET = 0
