### 2020-07-24 6f1ced7
Perform a clean_state after a regrid
