### 2020-07-24 e8e6abf
C++ doesn't like size 0 arrays for eos_t's aux (#1142)
block them out if NAUX_NET = 0
