+++ Source/driver/Castro.cpp
@@ -1427,14 +1427,14 @@ Castro::initialTimeStep ()
     }
     else
     {
-       init_dt = init_shrink*estTimeStep(dummy_dt);
+       init_dt = init_shrink*estTimeStep();
     }
 
     return init_dt;
 }
 
 Real
-Castro::estTimeStep (Real dt_old)
+Castro::estTimeStep ()
 {
     BL_PROFILE("Castro::estTimeStep()");
 
@@ -1609,14 +1609,14 @@ Castro::estTimeStep (Real dt_old)
 }
 
 void
-Castro::computeNewDt (int                   finest_level,
-                      int                   sub_cycle,
+Castro::computeNewDt (int                    finest_level,
+                      int                    /*sub_cycle*/,
                       Vector<int>&           n_cycle,
-                      const Vector<IntVect>& ref_ratio,
+                      const Vector<IntVect>& /*ref_ratio*/,
                       Vector<Real>&          dt_min,
                       Vector<Real>&          dt_level,
-                      Real                  stop_time,
-                      int                   post_regrid_flag)
+                      Real                   stop_time,
+                      int                    post_regrid_flag)
 {
     BL_PROFILE("Castro::computeNewDt()");
 
@@ -1632,7 +1632,7 @@ Castro::computeNewDt (int                   finest_level,
     for (int i = 0; i <= finest_level; i++)
     {
         Castro& adv_level = getLevel(i);
-        dt_min[i] = adv_level.estTimeStep(dt_level[i]);
+        dt_min[i] = adv_level.estTimeStep();
     }
 
     if (fixed_dt <= 0.0)
@@ -1820,9 +1820,9 @@ Castro::computeNewDt (int                   finest_level,
 
 void
 Castro::computeInitialDt (int                   finest_level,
-                          int                   sub_cycle,
+                          int                   /*subcycle*/,
                           Vector<int>&           n_cycle,
-                          const Vector<IntVect>& ref_ratio,
+                          const amrex::Vector<amrex::IntVect>& /*ref_ratio*/,
                           Vector<Real>&          dt_level,
                           Real                  stop_time)
 {
@@ -1867,13 +1867,13 @@ Castro::computeInitialDt (int                   finest_level,
 }
 
 void
-Castro::post_timestep (int iteration)
+Castro::post_timestep (int iteration_local)
 {
     BL_PROFILE("Castro::post_timestep()");
 
     // Pass some information about the state of the simulation to a Fortran module.
 
-    ca_set_amr_info(level, iteration, -1, -1.0, -1.0);
+    ca_set_amr_info(level, iteration_local, -1, -1.0, -1.0);
 
     //
     // Integration cycle on fine level grids is complete
@@ -1960,8 +1960,8 @@ Castro::post_timestep (int iteration)
 
         if (sum_per > 0.0) {
 
-          const int num_per_old = floor((cumtime - dtlev) / sum_per);
-          const int num_per_new = floor((cumtime        ) / sum_per);
+          const int num_per_old = static_cast<int>(std::floor((cumtime - dtlev) / sum_per));
+          const int num_per_new = static_cast<int>(std::floor((cumtime        ) / sum_per));
 
           if (num_per_old != num_per_new) {
             sum_per_test = true;
@@ -1995,9 +1995,9 @@ Castro::post_timestep (int iteration)
         //
         // Don't redistribute/timestamp on the final subiteration except on the coarsest grid.
         //
-        if (iteration < ncycle || level == 0)
[…]
