+++ Source/driver/Castro_generic_fill.H
@@ -29,11 +29,11 @@ static void clean_bc(int* bc_f)
 struct CastroGenericFill
 {
     AMREX_GPU_DEVICE
-    void operator() (const IntVect& iv, Array4<Real> const& data,
-                     const int dcomp, const int numcomp,
-                     GeometryData const& geom, const Real time,
-                     const BCRec* bcr, const int bcomp,
-                     const int orig_comp) const
+    void operator() (const IntVect& /*iv*/, Array4<Real> const& /*data*/,
+                     const int /*dcomp*/, const int /*numcomp*/,
+                     GeometryData const& /*geom*/, const Real /*time*/,
+                     const BCRec* /*bcr*/, const int /*bcomp*/,
+                     const int /*orig_comp*/) const
         {
             // Dummy routine that does nothing for inflow boundaries.
             // We assume that there are no inflow boundaries for a
