+++ .github/workflows/c-linter.yml
@@ -11,7 +11,7 @@ jobs:
         run: |
           git submodule update --init
           cd external/Microphysics
-          git fetch; git checkout development
+          git fetch; git checkout development 
           cd ../..
 
       - name: Run cpp linter
