+++ .github/workflows/c-linter.yml
@@ -12,6 +12,8 @@ jobs:
           git submodule update --init
           cd external/Microphysics
           git fetch; git checkout development
+          cd ../amrex 
+          git fetch; git checkout development
           cd ../..
 
       - name: Run cpp linter
