### 2020-07-27 #2513 @kegiljarhus
Dear @eloisabentivegna and @zingale. This is my first review for JOSS, so please bear with me. I have gone through the review checklist. I had no problems installing the software, running a problem and visualizing the results. Castro is a highly advanced software package which solves relevant scientific problems using state-of-the-art numerical methods. The software seems well-designed and the documentation is also extensive and well-written.

I only have a few minor comments to be addressed:

- Some developers listed as core developers in the preface of the documentation are not included as authors: Beckner, Day, Lijewski, Malone, Nonaka. There is also one author, Sexton, not listed as a core developer. Could you please comment on this?
- There appears to be some automated testing and regression testing, but it is not clear to me how these are executed as part of the continuous integration or how to run these manually. Could you please make this clearer, preferably in the README?
