### 2020-07-27 5a8df73
fix compilation due to aux no longer being available (#1149)
