### 2020-07-29 #2513 @eloisabentivegna
Hi @kegiljarhus, thanks for your review and update!

Could you raise those issues in Castro's repository? This will ensure your suggestions will enter the package's history and be properly credited. You can create a mention here by using this issue's URL in the repository's issue, if you wish (see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#guiding-principles).
