### 2020-07-29 #2513 @zingale
@kegiljarhus thank you for the comments.

For the first, it is indeed the case that those names were listed on the developers page, but none of them have made any contributions in the last 6 years and some of the names don't show in the git history.  According to our policy in the README.md, we consider developers as active if they've contributed in the last 3 years.  I've moved the inactive names to a "Former Developers" category (through this PR: https://github.com/AMReX-Astro/Castro/pull/1150 ).  I've also added Jean Sexton to the list of developers.

For the second, our testing is currently manually triggered on PRs using our regression testing framework here: https://github.com/AMReX-Codes/regression_testing .  Part of the reason is because we need to test against 3 different repos (Castro, Microphysics, and AMReX) and we want to make sure that not just Castro changes are clean, but that changes to those other repos also don't affect Castro.  Also, our test suite is expensive to run and also uses GPUs, so we use a local machine to do this.  For most PRs you'll see a link to the test results.  The master list of CPU tests is here: http://groot.astro.sunysb.edu/Castro/test-suite/gfortran/ and for GPUs is here: http://groot.astro.sunysb.edu/Castro/test-suite/gpu/
