### 2020-07-29 #2513 @zingale
Note: in addition to being triggered manually on PRs, the test suite also runs nightly using the latest AMReX, Castro, and Microphysics.
