+++ paper-parser/papers.bib
@@ -1019,18 +1019,22 @@ archivePrefix = "arXiv",
   subject = {Type Ia supernovae&mdash;mergers and collisions}
 }
 
-@ARTICLE{2019arXiv190412877C,
-   author = {{Chen}, K.-J. and {Woosley}, S.~E.},
-    title = "{Dynamical Impact of Ni Heating in the Pair-Instability Supernovae}",
-  journal = {arXiv e-prints},
-archivePrefix = "arXiv",
-   eprint = {1904.12877},
- primaryClass = "astro-ph.HE",
- keywords = {Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
-     year = 2019,
-    month = "apr",
-   adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190412877C},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+@ARTICLE{2020ApJ...897..152C,
+       author = {{Chen}, Ke-Jung and {Woosley}, S.~E. and {Whalen}, Daniel J.},
+        title = "{Gas Dynamics of the Nickel-56 Decay Heating in Pair-instability Supernovae}",
+      journal = {\apj},
+     keywords = {1664, 1129, 1668, 101, 1285, 775, 732, 100, 767, 2086, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2020,
+        month = "jul",
+       volume = {897},
+       number = {2},
+          eid = {152},
+        pages = {152},
+          doi = {10.3847/1538-4357/ab9819},
+archivePrefix = {arXiv},
+       eprint = {1904.12877},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...897..152C},
   subject = {pair-instability/pop III supernovae}
 }
 
