### 2020-07-30 #2513 @kegiljarhus
Thank you for the suggestions, @eloisabentivegna. Since @zingale already updated the developer list, and the pull request links to this review, I believe that is sufficient and we do not need to create a new issue. The comments on testing are also satisfactory from my side.
