### 2020-07-31 AMReX-Astro/Castro#1154 @bonh
JOSS: repo size
https://github.com/openjournals/joss-reviews/issues/2513

The directory `.git` has around 100 MB which is huge compared to the total size of the repo being around 130 MB, right? Perhaps there is some easy and convenient way to clean up the `.git` folder somehow? `git gc`?
