### 2020-07-31 AMReX-Astro/Castro#1155 @bonh
JOSS: list of requirements
https://github.com/openjournals/joss-reviews/issues/2513

I am not able to find a list of required packages. I can determine from [https://amrex-astro.github.io/Castro/docs/getting_started.html#building-the-code]() that at least I need a working compiler suite and perhaps a working MPI implementation. Is that all? Especially, if I plan to compile your code on a hpc cluster (you call it "Massively Parallel", so that seems to be the target platform, right?) an exhaustive list of requirements (and incompatibilities, e.g., compilers you know your code does not work with) is really helpful.

Secondly, it is not clear to me for what the `requirements.txt` in the root of the repo is used for.
