### 2020-07-31 #2513 @bonh
@eloisabentivegna On references: "Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)?" As this is not my field of expertise I am not able to determine if the references concerning the physics and mathematics for this particular problem are complete and/or appropriate.
