### 2020-07-31 #2513 @bonh
@eloisabentivegna On Functionality: "Have the functional claims of the software been confirmed?" I do not have a hpc cluster available for testing so I am not able to confirm the code to be able to conduct "massively parallel [...] simulations"
