### 2020-07-31 AMReX-Astro/Castro#1155 @bonh
Compiler flags you know work well would be great too.
