### 2020-07-31 AMReX-Astro/Castro#1155 @bonh
```--------------------------------------------------------------------------
The Open MPI wrapper compiler was unable to find the specified compiler
/usr/bin/gfortran in your PATH.

Note that this compiler was either specified at configure time or in
one of several possible environment variables.
--------------------------------------------------------------------------
make: *** [../../../external/amrex/Tools/GNUMake/Make.rules:246: tmp_build_dir/o/2d.gnu.MPI.EXE/AMReX_fort_mod.o] Error 1
```
So I do need a working fortran compiler, right?
