### 2020-07-31 AMReX-Astro/Castro#1154 @zingale
running `git gc` doesn't really change these sizes.  The history goes back 10 years, so I don't think that these relative sizes are a big issue, as the code has undergone a lot of changes in that decade.
