### 2020-07-31 AMReX-Astro/Castro#1155 @harpolea
Hi @bonh! The `requirements.txt` file is used in our docs build, as can be seen from looking at the action (https://github.com/AMReX-Astro/Castro/blob/development/.github/workflows/gh-pages.yml). We use Sphinx to build the documentation, and the `requirements.txt` file contains all the additional Python packages required.
