### 2020-07-31 AMReX-Astro/Castro#1155 @harpolea
The packages required are listed at the top of the Getting Started page that you linked to:
> Castro has two source dependencies: AMReX, the adaptive mesh library, and StarKiller Microphysics, the collection of equations of state, reaction networks, and other microphysics. The instructions below describe how to get these dependencies automatically with Castro.

If you follow the 'Downloading the code' section, then it will instruct you how to install these packages as git submodules.

We also suggest on that page several compilers (GNU, PGI and Intel), however we are updating it to make it more explicit that a C++ compiler, a Fortran 20xx compiler, python 3.x, and GNU make are required.
