### 2020-07-31 AMReX-Astro/Castro#1155 @harpolea
In addition to this, we have a page describing how to build our codes (including Castro) on several HPC clusters that we use in our group, as well as on a standard Linux workstation (https://amrex-astro.github.io/workflow/index.html). We shall be adding a link to this page on our Getting Started page.
