### 2020-07-31 AMReX-Astro/Castro#1155 @harpolea
Standard compiler flags are taken care of by the AMReX framework on which Castro is built (https://github.com/AMReX-Codes/amrex/tree/development/Tools/GNUMake/comps).

Castro defines a number of custom compiler flags (listed here https://amrex-astro.github.io/Castro/docs/build_system.html) which can be used to control the parallelism and which physics solvers are used.
