### 2020-07-31 fffd565
add some documentation about the build requirements (#1156)
This explains the build requirements in more detail, addressing some of #1155
