### 2020-07-31 AMReX-Astro/Castro#1155 @zingale
Here's a link to the updated documentation:

https://amrex-astro.github.io/Castro/docs/dev/getting_started.html
