### 2020-07-31 AMReX-Astro/Castro#1154 @bonh
Okay, I get this, thank you. But if you run `git rev-list --objects --all \
| git cat-file --batch-check='%(objecttype) %(objectname) %(objectsize) %(rest)' \
| sed -n 's/^blob //p' \
| sort --numeric-sort --key=2 \
| cut -c 1-12,41- \
| $(command -v gnumfmt || echo numfmt) --field=2 --to=iec-i --suffix=B --padding=7 --round=nearest`, which is the second answer from here [https://stackoverflow.com/questions/10622179/how-to-find-identify-large-commits-in-git-history](), you will notice that there are a bunch of larger blobs containing data or binaries, most likely pretty old (haven't checked). I assume, that you could remove the old data from the history and reduce the size of your repo significantly.

Note, that in my opinion this issue does not prevent acceptance by JOSS. So if you really do not think it is worth the effort feel free to close this issue.
