### 2020-07-31 AMReX-Astro/Castro#1154 @zingale
I'll take a look.

Part of the issue is that we report the git hashes of the code used in the science papers we write, so we really don't want to change the history.  And some of those older files are older initial models that have since been replaced, but if we want to go back to a science paper from a few years ago, we may need some of those.
