### 2020-07-31 b589eda
selective clang_format linting
