### 2020-07-31 97514cf
use https for cloning
