+++ .github/workflows/c-linter.yml
@@ -16,6 +16,11 @@ jobs:
           git fetch; git checkout development
           cd ../..
 
+      - name: install dependencies
+        run: |
+          apt-get -qq -y install curl clang-tidy cmake jq clang cppcheck clang-format bear g++>=9.3.0 gfortran>=9.3.0
+          apt-get update
+
       - name: Get cpp linter repo 
         run: |
           cd external 
