+++ .github/workflows/c-linter.yml
@@ -18,9 +18,8 @@ jobs:
 
       - name: install dependencies
         run: |
-          sudo apt-get update
+          sudo apt-get update -y -qq
           sudo apt-get -qq -y install curl clang-tidy cmake jq clang cppcheck clang-format bear g++>=9.3.0 gfortran>=9.3.0
-          apt-get update
 
       - name: Get cpp linter repo 
         run: |
