+++ docs/_sources/build_system.rst.txt
@@ -1,3 +1,5 @@
+.. _ch:buildsystem:
+
 *********************
 Build System Overview
 *********************
