+++ docs/_sources/getting_started.rst.txt
@@ -10,6 +10,25 @@ Getting Started
    instructions below describe how to get these dependencies automatically
    with Castro.
 
+
+The compilation process is managed by AMReX and its build system.  The
+general requirements to build Castro are:
+
+ * A C++ 11 (or later) compiler
+
+ * A Fortran 20xx compiler
+
+ * python (>= 3.6)
+
+ * GNU make (>= 3.81)
+
+GCC and the PGI compilers are the main compiler suites used by the
+developers.
+
+For running in parallel, an MPI library is required.  For running on GPUs,
+CUDA 10 or later is required.  More information on parallel builds
+is given in section :ref:`ch:mpiplusx`.
+
 Downloading the Code
 ====================
 
@@ -136,6 +155,8 @@ build the Sedov problem:
    The resulting executable will look something like
    ``Castro2d.gnu.ex``, which means this is a 2-d version
    of the code compiled with ``COMP = gnu``.
+
+More information on the various build options is given in :ref:`ch:buildsystem`.
 
 Running the Code
 ================
