### 2020-08-03 5ab1f32
increase fetch depth so it gets entire git history
