### 2020-08-04 4ea2813
turn off mpi and omp
