+++ .github/workflows/c-linter.yml
@@ -30,7 +30,7 @@ jobs:
           cd .. 
 
       - name: Run cpp linter
-        run: python3 external/cpp-linter-action/run_on_changed_files.py ${{ github.event.pull_request.base.sha }} ${{ github.event.pull_request.head.sha }}
+        run: python3 external/cpp-linter-action/run_on_changed_files.py ${{ github.event.pull_request.base.sha }} ${{ github.event.pull_request.head.sha }} -header-filter=Castro -ignore-files="amrex|Microphysics" 
       
       - name: Archive clang tidy report
         uses: actions/upload-artifact@v1
