### 2020-08-10 b30ef1b
check that timestep is not too small at time = 0
