+++ Source/driver/Castro_advance_ctu.cpp
@@ -515,7 +515,7 @@ Castro::subcycle_advance_ctu(const Real time, const Real dt, int amr_iteration, 
 
         // Determine whether we're below the cutoff timestep.
 
-        if (dt_subcycle < dt_cutoff * time) {
+        if (dt_subcycle < dt_cutoff * time || (time == 0.0 && dt_subcyle < eps)) {
             if (ParallelDescriptor::IOProcessor()) {
                 std::cout << std::endl;
                 std::cout << "  The subcycle mechanism requested subcycled timesteps of maximum length dt = " << dt_subcycle << "," << std::endl
