### 2020-08-12 #2513 @eloisabentivegna
@bonh, thanks for your questions about the reference list box -- please feel free to leave it unticked.

Regarding the code tests, we only really expect reviewers to check the small-scale examples (as a rough rule of thumb, those that can be executed on a reasonable laptop). For all the other tests, please simply check that the results appear sensible and generally consistent.

Thanks for all the useful comments!
