### 2020-08-19 3370c7c
Made mhd detonation problem.
