+++ Exec/mhd_tests/Detonation/bc_fill_nd.F90
@@ -1,0 +1,211 @@
+module bc_fill_module
+
+  use amrex_fort_module, only: rt => amrex_real
+
+  implicit none
+
+  public
+
+contains
+
+  ! Given a zone state, fill it with ambient material.
+
+  subroutine fill_ambient(state, s_lo, s_hi, i, j, k, x, time)
+
+    use amrex_constants_module, only: ZERO, HALF
+    use meth_params_module, only: NVAR, URHO, UMX, UMY, UMZ, UTEMP, UEINT, UEDEN, UFS
+    use network, only: nspec
+    use prob_params_module, only: problo, probhi
+    use probdata_module
+
+    implicit none
+
+    integer,  intent(in   ) :: s_lo(3), s_hi(3)
+    real(rt), intent(inout) :: state(s_lo(1):s_hi(1),s_lo(2):s_hi(2),s_lo(3):s_hi(3),NVAR)
+    integer,  intent(in   ) :: i, j, k
+    real(rt), intent(in   ) :: x, time
+
+    real(rt) :: c_T
+
+    !$gpu
+
+    c_T = problo(1) + center_T * (probhi(1) - problo(1))
+
+    state(i,j,k,URHO) = ambient_dens
+    state(i,j,k,UFS:UFS-1+nspec) = ambient_dens * ambient_comp
+
+    if (x < c_T) then
+       state(i,j,k,UTEMP) = T_l
+       state(i,j,k,UEINT) = state(i,j,k,URHO) * ambient_e_l
+       state(i,j,k,UMX  ) = state(i,j,k,URHO) * (vel + grav_acceleration * time)
+    else
+       state(i,j,k,UTEMP) = T_r
+       state(i,j,k,UEINT) = state(i,j,k,URHO) * ambient_e_r
+       state(i,j,k,UMX  ) = -state(i,j,k,URHO) * (vel + grav_acceleration * time)
+    end if
+
+    state(i,j,k,UMY:UMZ) = ZERO
+    state(i,j,k,UEDEN) = state(i,j,k,UEINT) + HALF * sum(state(i,j,k,UMX:UMZ)**2) / state(i,j,k,URHO)
+
+  end subroutine fill_ambient
+
+
+
+  subroutine hypfill(lo, hi, adv, adv_lo, adv_hi, domlo, domhi, delta, xlo, time, bc) bind(C, name="hypfill")
+
+    use amrex_constants_module, only: HALF
+    use meth_params_module, only: NVAR, fill_ambient_bc
+    use prob_params_module, only : problo
+
+    implicit none
+
+    include 'AMReX_bc_types.fi'
+
+    integer,  intent(in   ) :: lo(3), hi(3)
+    integer,  intent(in   ) :: adv_lo(3), adv_hi(3)
+    integer,  intent(in   ) :: bc(AMREX_SPACEDIM, 2, NVAR)
+    integer,  intent(in   ) :: domlo(3), domhi(3)
+    real(rt), intent(in   ) :: delta(3), xlo(3)
+    real(rt), intent(inout) :: adv(adv_lo(1):adv_hi(1),adv_lo(2):adv_hi(2),adv_lo(3):adv_hi(3),NVAR)
+    real(rt), intent(in   ), value :: time
+
+    integer  :: i, j, k, n
+    real(rt) :: x
+
+    !$gpu
+
+    ! Override the generic routine at the physical boundaries by
+    ! setting the material to the ambient state. Note that we
+    ! don't want to do this for interior/periodic boundaries,
+    ! which have bc == 0, or for reflecting boundaries, which have
+    ! bc == -1 or bc == 1.
+
+    if (fill_ambient_bc == 1) then
+
[…]
