### 2020-08-20 235adc9
sync this up to the new set of parameters for a aprox19 det this disabled the NSE table at the moment
