### 2020-08-20 f81f896
Edits made to mhd detonations problem
