### 2020-08-20 dfdec06
output if a run crashed, fix some spacing
