### 2020-08-21 83e3bce
if we are doing NSE, then set the aux quantities (#1166)
