### 2020-08-21 310faf9
Updated MHD Detonation problem
