### 2020-08-22 ff8f102
better initial conditions -- this seems to resolve the shock ignition it also does the retry of the burn better
