### 2020-08-22 1d67f73
fix crash check
