### 2020-08-22 4b39313
split out the detonation class
