### 2020-08-22 d8e49a4
use the detonation module
