### 2020-08-22 be932f8
add in the 3 iteration plots
