### 2020-08-24 0b451d8
Compiler warnings 2 (#1145)
address more compiler warnings
