+++ Source/driver/Castro_nd.F90
@@ -87,23 +87,6 @@ end subroutine ca_extern_init
 
 
 #ifdef REACTIONS
-subroutine ca_get_abort_on_failure(abort_on_failure_in) bind(C, name="ca_get_abort_on_failure")
-
-  use extern_probin_module, only : abort_on_failure
-
-  implicit none
-
-  integer, intent(inout) :: abort_on_failure_in
-
-  if (abort_on_failure) then
-     abort_on_failure_in = 1
-  else
-     abort_on_failure_in = 0
-  endif
-
-end subroutine ca_get_abort_on_failure
-
-
 subroutine ca_set_abort_on_failure(abort_on_failure_in) bind(C, name="ca_set_abort_on_failure")
 
   use extern_probin_module, only : abort_on_failure
