### 2020-08-25 67f7bfe
address more compiler warnings
