+++ Source/driver/Castro_advance_sdc.cpp
@@ -110,7 +110,7 @@ Castro::do_advance_sdc (Real time,
           }
 
           // we pass in the stage time here
-          do_old_sources(old_source, Sburn, Sburn, node_time, dt, apply_sources_to_state, amr_iteration, amr_ncycle);
+          do_old_sources(old_source, Sburn, Sburn, node_time, dt, apply_sources_to_state);
 
           // fill the ghost cells for the sources -- note since we have
           // not defined the new_source yet, we either need to copy this
@@ -136,7 +136,7 @@ Castro::do_advance_sdc (Real time,
         } else {
           // there is a ghost cell fill hidden in diffusion, so we need
           // to pass in the time associate with Sborder
-          do_old_sources(old_source, Sborder, Sborder, cur_time, dt, apply_sources_to_state, amr_iteration, amr_ncycle);
+          do_old_sources(old_source, Sborder, Sborder, cur_time, dt, apply_sources_to_state);
         }
 
         // note: we don't need a FillPatch on the sources, since they
@@ -274,12 +274,12 @@ Castro::do_advance_sdc (Real time,
     // TODO: we also need to make these 4th order!
     clean_state(S_old, prev_time, 0);
     expand_state(Sborder, prev_time, Sborder.nGrow());
-    do_old_sources(old_source, Sborder, Sborder, prev_time, dt, apply_sources_to_state, amr_iteration, amr_ncycle);
+    do_old_sources(old_source, Sborder, Sborder, prev_time, dt, apply_sources_to_state);
     AmrLevel::FillPatch(*this, old_source, old_source.nGrow(), prev_time, Source_Type, 0, NSRC);
 
     clean_state(S_new, cur_time, 0);
     expand_state(Sborder, cur_time, Sborder.nGrow());
-    do_old_sources(new_source, Sborder, Sborder, cur_time, dt, apply_sources_to_state, amr_iteration, amr_ncycle);
+    do_old_sources(new_source, Sborder, Sborder, cur_time, dt, apply_sources_to_state);
     AmrLevel::FillPatch(*this, new_source, new_source.nGrow(), cur_time, Source_Type, 0, NSRC);
   }
 
