### 2020-08-25 d4aafd2
fix a bunch more compiler warnings (#1175)
