### 2020-08-30 5a03ad0
fix the clean_state calls when we do MHD + reactions also make the reacting_convergence inputs work for 3-d
