+++ Docs/source/FlowChart.rst
@@ -35,7 +35,7 @@ the different code paths.  These fall into two categories:
      time using a simple quadrature rule, and integrates the hydro
      explicitly and reactions implicitly to the next time node.
      Iterations allow each process to see one another and achieve
-     high-order in time convergence.  This is described in :cite:`castro_sdc`.
+     high-order in time convergence.  This is described in :cite:`castro-sdc`.
 
 
 The time-integration method used is controlled by
@@ -63,6 +63,18 @@ The time-integration method used is controlled by
     state variables, you must compile with ``USE_SIMPLIFIED_SDC = TRUE`` for this
     method to work (in particular, this defines ``PRIM_SPECIES_HAVE_SOURCES``).
 
+.. index:: USE_SIMPLIFIED_SDC, USE_TRUE_SDC
+
+.. note::
+
+   By default, the code is compiled for Strang-split CTU evolution
+   (``time_integration_method = 0``).  Because the size of the
+   different state arrays differs with the other integration schemes,
+   support for them needs to be compiled in, using
+   ``USE_SIMPLIFIED_SDC=TRUE`` for the simplified-SDC method
+   (``time_integration_method=3``) and ``USE_TRUE_SDC=TRUE`` for the
+   true SDC method (``time_integration_method = 2``).
+
 Several helper functions are used throughout:
 
 .. index:: clean_state
@@ -74,7 +86,11 @@ Several helper functions are used throughout:
 
    #. enforces that the density is above ``castro.small_dens``
 
+   #. enforces that the speeds in the state don't exceed ``castro.speed_limit``
+
    #. normalizes the species so that the mass fractions sum to 1
+
+   #. syncs up the linear and hybrid momenta (for ``USE_HYBRID_MOMENTUM=TRUE``)
 
    #. resets the internal energy if necessary (too small or negative)
       and computes the temperature for all zones to be thermodynamically
@@ -492,25 +508,28 @@ In the code, the objective is to evolve the state from the old time,
 A summary of which state is the input and which is updated for each of
 these processes is presented below:
 
-+--------------------+-----------+---------------------+---------------------+
-| *step*             | ``S_old`` | ``Sborder``         | ``S_new``           |
-+====================+===========+=====================+=====================+
-| 1. init            | input     | updated             |                     |
-+--------------------+-----------+---------------------+---------------------+
-| 2. react           |           | input / updated     |                     |
-+--------------------+-----------+---------------------+---------------------+
-| 3. old sources     |           | input               | updated             |
-+--------------------+-----------+---------------------+---------------------+
-| 4. hydro           |           | input               | updated             |
-+--------------------+-----------+---------------------+---------------------+
-| 5. clean           |           |                     | input / updated     |
-+--------------------+-----------+---------------------+---------------------+
-| 6. radial / center |           |                     | input               |
-+--------------------+-----------+---------------------+---------------------+
-| 7. correct sources |           |                     | input / updated     |
-+--------------------+-----------+---------------------+---------------------+
-| 8. react           |           |                     | input / updated     |
-+--------------------+-----------+---------------------+---------------------+
+.. table:: update sequence of state arrays for Strang-CTU
+   :align: center
+
+   +--------------------+-----------+---------------------+---------------------+
+   | *step*             | ``S_old`` | ``Sborder``         | ``S_new``           |
+   +====================+===========+=====================+=====================+
+   | 1. init            | input     | updated             |                     |
+   +--------------------+-----------+---------------------+---------------------+
+   | 2. react           |           | input / updated     |                     |
[…]
