### 2020-09-01 5e3db41
Fixes bug which caused fourth order SDC to crash when run with OMP
