### 2020-09-02 0742517
remove references to unused rpar parameters (#1182)
We removed the t_sound based burning limiter, this now removes references to those parameters. This is needed
once we merge starkiller-astro/Microphysics#388
