### 2020-09-02 1d8dd4c
only compile in the conservative_interpolate routines if we ask for them also make scf wrapped with GRAVITY
