### 2020-09-02 adbed88
we don't use the Fortran version of mom_flux_has_p anymore
