### 2020-09-03 d1286b2
start converting to C++
