### 2020-09-03 3da1748
start converting the radiation consup to C++
