+++ Source/radiation/RadHydro.cpp
@@ -1,0 +1,286 @@
+
+constexpr int rk_order = 3;
+constexpr bool use_WENO = false;
+
+constexpr Real cfl = 0.5_rt;
+
+constexpr Real onethird = 1.0_rt/3.0_rt;
+constexpr Real twothirds = 2.0_rt/3.0_rt;
+constexpr Real onesixth = 1.0_rt/6.0_rt;
+
+// RK5
+constexpr Real B1 = 0.5_rt;
+constexpr Real B2 = 1.0_rt/16.0_rt;
+constexpr Real B3 = 0.5_rt;
+constexpr Real B4 = 9.0_rt/16.0_rt;
+constexpr Real B5 = 8.0_rt/7.0_rt;
+constexpr Real B6 = 7.0_rt/90.0_rt;
+
+constexpr Real C20 = 5.0_rt/8.0_rt;
+constexpr Real C21 = 3.0_rt/8.0_rt;
+
+constexpr Real C40 = 17.0_rt/8.0_rt;
+constexpr Real C41 = 9.0_rt/8.0_rt;
+constexpr Real C42 = -3.0_rt;
+constexpr Real C43 = 0.75_rt;
+
+constexpr Real C50 = -5.0_rt/21.0_rt;
+constexpr Real C51 = 2.0_rt/7.0_rt;
+constexpr Real C52 = 0.0_rt;
+constexpr Real C53 = 4.0_rt;
+constexpr Real C54 = -64.0_rt/21.0_rt;
+
+constexpr Real C60 = -8.0_rt/27.0_rt;
+constexpr Real C61 = -1.0_rt/5.0_rt;
+constexpr Real C62 = 32.0_rt/45.0_rt;
+constexpr Real C63 = -32.0_rt/45.0_rt;
+constexpr Real C64 = 32.0_rt/27.0_rt;
+constexpr Real C65 = 14.0_rt/45.0_rt;
+
+
+void advect_in_fspace(Real* ustar, Real* af, const Real dt, Real& nstep_fsp) {
+
+  call update_one_species(Radiation::nGroups, ustar, af, dlognu, dt, nstep_fsp)
+
+    else
+
+       call update_one_species(ng0, ustar(0:ng0-1), &
+                               af(0:ng0-1), &
+                               dlognu(0:ng0-1), &
+                               dt, nstep_fsp)
+
+       if (nnuspec >= 2) then
+          call update_one_species(ng1, ustar(ng0:ng0+ng1-1), &
+                                  af(ng0:ng0+ng1-1), &
+                                  dlognu(ng0:ng0+ng1-1), &
+                                  dt, nstep_fsp)
+       end if
+
+       if (nnuspec == 3) then
+          ng2 = ngroups-ng0-ng1
+          call update_one_species(ng2, ustar(ng0+ng1:), &
+                                  af(ng0+ng1:), &
+                                  dlognu(ng0+ng1:), &
+                                  dt, nstep_fsp)
+       end if
+
+    end if ! end of if nnuspec
+
+  end subroutine advect_in_fspace
+
+
+  subroutine update_one_species(n, u, a, dx, tend, nstepmax)
+
+    use amrex_fort_module, only: rt => amrex_real
+
+    implicit none
+
+    integer, intent(in) :: n
[…]
