+++ Source/hydro/Castro_ctu_rad.cpp
@@ -1,0 +1,300 @@
+#include "Castro.H"
+#include "Castro_F.H"
+#include "Castro_util.H"
+#include "Castro_hydro_F.H"
+
+#ifdef RADIATION
+#include "Radiation.H"
+#endif
+
+using namespace amrex;
+
+#ifdef RADIATION
+Castro::ctu_rad_consup(const Box& bx,
+                       Array4<Real> const& update,
+                       Array4<Real const> const& Erin,
+                       Array4<Real> const& uout,
+                       Array4<Real> const& Erout,
+                       Array4<Real const> const& radflux1,
+                       Array4<Real const> const& qx,
+                       Array4<Real const> const& area1,
+#if AMREX_SPACEDIM >= 2
+                       Array4<Real const> const& radflux2,
+                       Array4<Real const> const& qy,
+                       Array4<Real const> const& area2,
+#endif
+#if AMREX_SPACEDIM == 3
+                       Array4<Real const> const& radflux3,
+                       Array4<Real const> const& qz,
+                       Array4<Real const> const& area3,
+#endif
+                       Real& nstep_fsp,
+                       Array4<Real const> const& vol,
+                       const Real dt)
+{
+
+  Real Erscale = 0;
+  if (ngroups > 1) {
+    if (fspace_type == 1) {
+      Erscale = dlognu;
+    } else {
+      Erscale = nugroup * dlognu;
+    }
+  }
+
+  int fspace_t = Radiation::fspace_advection_type;
+  int comov = Radiation::comoving;
+  int limiter = Radiation::limiter;
+  int closure = Radiation::closure;
+
+  // radiation energy update.  For the moment, we actually update things
+  // fully here, instead of creating a source term for the update
+  amrex::ParallelFor(bx, Radiation::nGroups,
+  [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k, int g) noexcept
+  {
+
+    Erout(i,j,k,g) = Erin(i,j,k,g) + dt *
+      (  radflux1(i,j,k,g) * area1(i,j,k) - radflux1(i+1,j,k,g) * area1(i+1,j,k)
+#if AMREX_SPACEDIM >= 2
+       + radflux2(i,j,k,g) * area2(i,j,k) - radflux2(i,j+1,k,g) * area2(i,j+1,k) +
+#endif
+#if AMREX_SPACEDIM == 3
+       + radflux3(i,j,k,g) * area3(i,j,k) - radflux3(i,j,k+1,g) * area3(i,j,k+1)
+#endif
+         ) / vol(i,j,k);
+  });
+
+  // Add gradp term to momentum equation -- only for axisymmetry coords
+  // (and only for the radial flux);  also add the radiation pressure gradient
+  // to the momentum for all directions
+  amrex::ParallelFor(bx,
[…]
