### 2020-09-03 40ef024
more work on fspace
