+++ Source/radiation/RadHydro.cpp
@@ -1,4 +1,7 @@
 
+// do the advection in frequency space for the radiation energy.
+// see Paper III, section 2.4
+
 constexpr int rk_order = 3;
 constexpr bool use_WENO = false;
 
@@ -40,117 +43,119 @@ constexpr Real C65 = 14.0_rt/45.0_rt;
 
 void advect_in_fspace(Real* ustar, Real* af, const Real dt, Real& nstep_fsp) {
 
-  call update_one_species(Radiation::nGroups, ustar, af, dlognu, dt, nstep_fsp)
+  call update_one_species(nGroups, ustar, af, dlognu, dt, nstep_fsp)
 
-    else
+}
 
-       call update_one_species(ng0, ustar(0:ng0-1), &
-                               af(0:ng0-1), &
-                               dlognu(0:ng0-1), &
-                               dt, nstep_fsp)
+void update_one_species(const int n, Real* u, Real* a, Real* dx, const Real tend, int& nstepmax) {
 
-       if (nnuspec >= 2) then
-          call update_one_species(ng1, ustar(ng0:ng0+ng1-1), &
-                                  af(ng0:ng0+ng1-1), &
-                                  dlognu(ng0:ng0+ng1-1), &
-                                  dt, nstep_fsp)
-       end if
+  Real dt = 1.e50_rt;
+  Real acfl;
 
-       if (nnuspec == 3) then
-          ng2 = ngroups-ng0-ng1
-          call update_one_species(ng2, ustar(ng0+ng1:), &
-                                  af(ng0+ng1:), &
-                                  dlognu(ng0+ng1:), &
-                                  dt, nstep_fsp)
-       end if
+  int nstep;
 
-    end if ! end of if nnuspec
+  for (int i = 0; i < n; i++) {
+    acfl = 1.e-50_rt + std::abs(a[i]);
+    dt = amrex::min(dt, dx[i]/acfl*cfl);
+  }
 
-  end subroutine advect_in_fspace
+  if (dt >= tend) {
+    nstep = 1;
+    dt = tend;
+  } else {
+    nstep = std::ceil(tend/dt);
+    dt = tend / static_cast<Real>(nstep);
+  }
+
+  Real u1[n];
+  Real u2[n];
+  Real u3[n];
+  Real u4[n];
+  Real u5[n];
+  Real dudt_tmp[n];
+
+  for (int istep = 1; istep < nstep; istep++) {
+    if (rk_order == 5) {
+      // RK5
+      dudt(u, a, dx, n, dudt_tmp);
+      for (int g = 0; g < n; g++) {
+        u1[g] = u[g] + B1 * dt * dudt_tmp[g];
+      }
+      dudt(u1, a, dx, n, dudt_tmp);
+      for (int g = 0; g < n; g++) {
+        u2[g] = (C20*u[g] + C21*u1[g]) + B2 * dt * dudt_tmp[g];
+      }
+      dudt(u2, a, dx, n, dudt_tmp);
+      for (int g = 0; g < n; g++) {
+        u3[g] = u[g] + B3 * dt * dudt_tmp[g];
+      }
+      dudt(u3, a, dx, n, dudt_tmp);
+      for (int g = 0; g < n; g++) {
+        u4[g] = (C40*u[g] + C41*u1[g] + C42*u2[g] + C43*u3[g]) + B4 * dt * dudt_tmp[g];
+      }
[…]
