+++ Source/radiation/RadHydro_nd.F90
@@ -4,263 +4,7 @@ module radhydro_nd_module
 
   implicit none
 
-  integer, parameter :: rk_order = 3
-  logical, parameter :: use_WENO = .false.
-
-  real(rt), parameter :: cfl = 0.5e0_rt
-
-  real(rt), parameter :: onethird=1.e0_rt/3.e0_rt, twothirds=2.e0_rt/3.e0_rt, onesixth=1.e0_rt/6.e0_rt
-
-  ! RK5
-  real(rt), parameter :: B1=0.5e0_rt, B2=1.e0_rt/16.e0_rt, B3=0.5e0_rt, B4=9.e0_rt/16.e0_rt, &
-                         B5=8.e0_rt/7.e0_rt, B6=7.e0_rt/90.e0_rt
-  real(rt), parameter :: C20=5.e0_rt/8.e0_rt , C21=3.e0_rt/8.e0_rt
-  real(rt), parameter :: C40=17.e0_rt/8.e0_rt , C41=9.e0_rt/8.e0_rt , C42=-3.e0_rt , C43=0.75e0_rt
-  real(rt), parameter :: C50=-5.e0_rt/21.e0_rt , C51=2.e0_rt/7.e0_rt , C52=0.e0_rt , &
-                         C53=4.e0_rt , C54=-64.e0_rt/21.e0_rt
-  real(rt), parameter :: C60=-8.e0_rt/27.e0_rt , C61=-1.e0_rt/5.e0_rt , C62=32.e0_rt/45.e0_rt , &
-                         C63=-32.e0_rt/45.e0_rt, C64=32.e0_rt/27.e0_rt , C65=14.e0_rt/45.e0_rt
-
 contains
-
-  subroutine advect_in_fspace(ustar, af, dt, nstep_fsp)
-
-    use rad_params_module, only: ngroups, ng0, ng1, dlognu
-    use amrex_fort_module, only: rt => amrex_real
-
-    implicit none
-
-    real(rt), intent(inout) :: ustar(0:ngroups-1)
-    real(rt), intent(in) :: af(0:ngroups-1)
-    real(rt), intent(in) :: dt
-    integer, intent(inout) :: nstep_fsp
-    integer :: ng2
-
-    !$gpu
-
-    call update_one_species(ngroups, ustar, af, dlognu, dt, nstep_fsp)
-
-    end if
-
-  end subroutine advect_in_fspace
-
-
-  subroutine update_one_species(n, u, a, dx, tend, nstepmax)
-
-    use amrex_fort_module, only: rt => amrex_real
-
-    implicit none
-
-    integer, intent(in) :: n
-    real(rt), intent(inout) :: u(0:n-1)
-    real(rt), intent(in) :: a(0:n-1), dx(0:n-1)
-    real(rt), intent(in) :: tend
-    integer, intent(inout) :: nstepmax
-
-    real(rt) :: dt, acfl
-    real(rt) :: f(0:n), u1(0:n-1), u2(0:n-1), u3(0:n-1), u4(0:n-1), u5(0:n-1)
-    integer :: i, istep, nstep
-
-    !$gpu
-
-    dt = 1.e50_rt
-    do i=0, n-1
-       acfl = 1.e-50_rt + abs(a(i))
-       dt = min(dt, dx(i)/acfl*cfl)
-    end do
-
-    if (dt >= tend) then
-       nstep = 1
-       dt = tend
-    else
-       nstep = ceiling(tend/dt)
[…]
