### 2020-09-03 111cfe8
remove a bunch on unused functions
