+++ Source/radiation/Rad_nd.F90
@@ -2586,21 +2586,6 @@ end subroutine ca_initgroups3
 
 !! -----------------------------------------------------------
 
-subroutine ca_setgroup(igroup)
-
-  use rad_params_module, only: current_group
-  use amrex_fort_module, only: rt => amrex_real
-
-  implicit none
-
-  integer :: igroup
-
-  current_group = igroup
-
-end subroutine ca_setgroup
-
-!! -----------------------------------------------------------
-
 subroutine ca_inelastic_sct(lo, hi, &
                             uu,uu_lo,uu_hi, &
                             Er,Er_lo,Er_hi, &
