+++ Source/radiation/RAD_F.H
@@ -39,6 +39,9 @@ void ca_initsinglegroup
 void ca_get_dlognu
   (amrex::Real* dlognu_out);
 
+void ca_get_nugroup
+  (amrex::Real* nugroup);
+
 #ifdef __cplusplus
 }
 #endif
