+++ Source/hydro/flatten.cpp
@@ -1,6 +1,5 @@
 #include "Castro.H"
 #include "Castro_F.H"
-#include "Castro_hydro_F.H"
 #include <cmath>
 
 #ifdef RADIATION
