+++ Source/driver/Castro_F.H
@@ -200,9 +200,6 @@ extern "C"
 
 
 
-/* hydrodynamics */
-#include <Castro_hydro_F.H>
-
 
 /* reactions */
 #ifdef REACTIONS
