+++ Source/hydro/Castro_ctu_rad.cpp
@@ -1,7 +1,7 @@
 #include "Castro.H"
 #include "Castro_F.H"
 #include "Castro_util.H"
-#include "Castro_hydro_F.H"
+#include "Castro_hydro.H"
 
 #include "Radiation.H"
 #include "RadHydro.H"
