+++ Source/hydro/Castro_ctu_hydro.cpp
@@ -1274,27 +1274,27 @@ Castro::construct_ctu_hydro_source(Real time, Real dt)
 
 #ifdef RADIATION
 #pragma gpu box(bx)
-      ctu_rad_consup(AMREX_INT_ANYD(bx.loVect()), AMREX_INT_ANYD(bx.hiVect()),
-                     BL_TO_FORTRAN_ANYD(hydro_source[mfi]),
-                     BL_TO_FORTRAN_ANYD(Erborder[mfi]),
-                     BL_TO_FORTRAN_ANYD(S_new[mfi]),
-                     BL_TO_FORTRAN_ANYD(Er_new[mfi]),
-                     BL_TO_FORTRAN_ANYD(rad_flux[0]),
-                     BL_TO_FORTRAN_ANYD(qe[0]),
-                     BL_TO_FORTRAN_ANYD(area[0][mfi]),
+      ctu_rad_consup(bx,
+                     update_arr,
+                     Erborder.array(mfi),
+                     S_new.array(mfi),
+                     Er_new.array(mfi),
+                     (rad_flux[0]).array(),
+                     (qe[0]).array(),
+                     (area[0]).array(mfi),
 #if AMREX_SPACEDIM >= 2
-                     BL_TO_FORTRAN_ANYD(rad_flux[1]),
-                     BL_TO_FORTRAN_ANYD(qe[1]),
-                     BL_TO_FORTRAN_ANYD(area[1][mfi]),
+                     (rad_flux[1]).array(),
+                     (qe[1]).array(),
+                     (area[1]).array(mfi),
 #endif
 #if AMREX_SPACEDIM == 3
-                     BL_TO_FORTRAN_ANYD(rad_flux[2]),
-                     BL_TO_FORTRAN_ANYD(qe[2]),
-                     BL_TO_FORTRAN_ANYD(area[2][mfi]),
+                     (rad_flux[2]).array(),
+                     (qe[2]).array(),
+                     (area[2]).array(mfi),
 #endif
-                     &priv_nstep_fsp,
-                     BL_TO_FORTRAN_ANYD(volume[mfi]),
-                     AMREX_REAL_ANYD(dx), dt);
+                     priv_nstep_fsp,
+                     volume.array(mfi),
+                     dt);
 
       nstep_fsp = std::max(nstep_fsp, priv_nstep_fsp);
 #endif
