+++ Source/radiation/RadHydro.H
@@ -97,7 +97,7 @@ weno5(const Real vm2, const Real vm1, const Real v, const Real vp1, const Real v
 
 AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
 void
-dudt(Real* u, Real* a, Real* dx, const int n, Real* dudt_tmp) {
+dudt(Real* u, Real* a, const Real* dx, const int n, Real* dudt_tmp) {
 
   // compute the RHS for the g'th energy group out of n note, since
   // this is advection in frequency space, we will need to assume some
@@ -193,7 +193,7 @@ dudt(Real* u, Real* a, Real* dx, const int n, Real* dudt_tmp) {
 
 AMREX_GPU_HOST_DEVICE AMREX_INLINE
 void
-update_one_species(const int n, Real* u, Real* a, Real* dx, const Real tend, int& nstepmax) {
+update_one_species(const int n, Real* u, Real* a, const Real* dx, const Real tend, int& nstepmax) {
 
   Real dt = 1.e50_rt;
   Real acfl;
