### 2020-09-04 3f831bf
dlognu and nugroup are only defined for NGROUP > 1
