### 2020-09-04 1b8fccb
remove the unused nnuspec (#1188)
we no longer have neutrino support
