+++ Source/radiation/RadPlotvar_3d.f90
@@ -5,7 +5,7 @@ subroutine ca_er_com2lab(lo, hi, &
      F,     F_l1,  F_l2,  F_l3,  F_h1,  F_h2,  F_h3, iflx, nflx, & 
      Elab, El_l1, El_l2, El_l3, El_h1, El_h2, El_h3, ier, npv)
   use meth_params_module, only : NVAR, URHO, UMX, UMY, UMZ
-  use rad_params_module, only : ngroups, clight, nnuspec, ng0, ng1, dlognu
+  use rad_params_module, only : ngroups, clight, ng0, ng1, dlognu
   use amrex_fort_module, only : rt => amrex_real
   implicit none
 
@@ -47,91 +47,23 @@ subroutine ca_er_com2lab(lo, hi, &
      end do
      
      if (ngroups > 1) then
-        if (nnuspec .eq. 0) then
-           
-           do g=0,ngroups-1
-              nufnux(g) = F(i,j,k,ifx+g)*dlognuInv(g)
-              nufnuy(g) = F(i,j,k,ify+g)*dlognuInv(g)
-              nufnuz(g) = F(i,j,k,ifz+g)*dlognuInv(g)
-           end do
-           nufnux(-1) = -nufnux(0)
-           nufnuy(-1) = -nufnuy(0)
-           nufnuz(-1) = -nufnuz(0)
-           nufnux(ngroups) = -nufnux(ngroups-1)
-           nufnuy(ngroups) = -nufnuy(ngroups-1)              
-           nufnuz(ngroups) = -nufnuz(ngroups-1)              
-           do g=0,ngroups-1
-              Elab(i,j,k,g+ier) = Elab(i,j,k,g+ier) &
-                   - vxc2*0.5e0_rt*(nufnux(g+1)-nufnux(g-1)) &
-                   - vyc2*0.5e0_rt*(nufnuy(g+1)-nufnuy(g-1)) &
-                   - vzc2*0.5e0_rt*(nufnuz(g+1)-nufnuz(g-1))
-           end do
-           
-        else
-           
-           do g=0,ng0-1
-              nufnux(g) = F(i,j,k,ifx+g)*dlognuInv(g)
-              nufnuy(g) = F(i,j,k,ify+g)*dlognuInv(g)
-              nufnuz(g) = F(i,j,k,ifz+g)*dlognuInv(g)
-           end do
-           nufnux(-1) = -nufnux(0)
-           nufnuy(-1) = -nufnuy(0)
-           nufnuz(-1) = -nufnuz(0)
-           nufnux(ng0) = -nufnux(ng0-1)
-           nufnuy(ng0) = -nufnuy(ng0-1)              
-           nufnuz(ng0) = -nufnuz(ng0-1)              
-           do g=0,ng0-1
-              Elab(i,j,k,g+ier) = Elab(i,j,k,g+ier) &
-                   - vxc2*0.5e0_rt*(nufnux(g+1)-nufnux(g-1)) &
-                   - vyc2*0.5e0_rt*(nufnuy(g+1)-nufnuy(g-1)) &
-                   - vzc2*0.5e0_rt*(nufnuz(g+1)-nufnuz(g-1))
-           end do
-           
-           if (nnuspec >= 2) then
-              
-              do g=ng0,ng0+ng1-1
-                 nufnux(g) = F(i,j,k,ifx+g)*dlognuInv(g)
[…]
