### 2020-09-04 70b2e72
remove a bunch on unused functions (#1190)
