+++ Source/hydro/Castro_ctu_rad.cpp
@@ -77,7 +77,7 @@ Castro::ctu_rad_consup(const Box& bx,
     Erout(i,j,k,g) = Erin(i,j,k,g) + dt *
       (  radflux1(i,j,k,g) * area1(i,j,k) - radflux1(i+1,j,k,g) * area1(i+1,j,k)
 #if AMREX_SPACEDIM >= 2
-       + radflux2(i,j,k,g) * area2(i,j,k) - radflux2(i,j+1,k,g) * area2(i,j+1,k) +
+       + radflux2(i,j,k,g) * area2(i,j,k) - radflux2(i,j+1,k,g) * area2(i,j+1,k)
 #endif
 #if AMREX_SPACEDIM == 3
        + radflux3(i,j,k,g) * area3(i,j,k) - radflux3(i,j,k+1,g) * area3(i,j,k+1)
