### 2020-09-04 6c65570
fix a bounds issue in ca_initgroups3
