### 2020-09-04 6827ea1
fix a bounds issue in ca_initgroups3 (#1192)
In RadMultiGroup.cpp we are allocating xnu as size nGroups+1, but in the Fortran, we dimensioned it as 0:ngroups+2.
