+++ Source/driver/Castro_nd.F90
@@ -1,68 +1,3 @@
-
-
-subroutine ca_network_init() bind(C, name="ca_network_init")
-    !
-    ! Binds to C function `ca_network_init`
-    !
-
-  use network, only: network_init
-#ifdef REACTIONS
-  use actual_rhs_module, only: actual_rhs_init
-#endif
-
-  call network_init()
-
-#ifdef REACTIONS
-  call actual_rhs_init()
-#endif
-
-end subroutine ca_network_init
-
-
-subroutine ca_network_finalize() bind(C, name="ca_network_finalize")
-    !
-    ! Binds to C function `ca_network_finalize`
-    !
-
-  use network, only: network_finalize
-
-  call network_finalize()
-
-end subroutine ca_network_finalize
-
-
-subroutine ca_eos_init() bind(C, name="ca_eos_init")
-
-  use meth_params_module, only: small_dens, small_temp
-  use eos_module, only: eos_init
-
-  implicit none
-
-  call eos_init(small_dens=small_dens, small_temp=small_temp)
-
-  !$acc update device(small_dens, small_temp)
-
-end subroutine ca_eos_init
-
-
-subroutine ca_eos_finalize() bind(C, name="ca_eos_finalize")
-    !
-    ! Binds to C function `ca_eos_finalize`
-    !
-
-  use eos_module, only: eos_finalize
-
-  call eos_finalize()
-
-end subroutine ca_eos_finalize
-
-
-! :::
-! ::: ----------------------------------------------------------------
-! :::
-
-
-
 subroutine ca_extern_init(name,namlen) bind(C, name="ca_extern_init")
     ! initialize the external runtime parameters in
     ! extern_probin_module
