+++ Source/driver/Castro_F.H
@@ -9,16 +9,6 @@ extern "C"
 {
 #endif
 
-  void ca_network_init();
-
-  void ca_network_finalize();
-
-  void ca_eos_init();
-
-  void ca_eos_finalize();
-
-  void burner_init();
-
   void ca_extern_init(const int* name, const int* namlen);
 
 #ifdef REACTIONS
