### 2020-09-05 f87c832
remove some unused funcs + move ca_normalize_species into SDC
