+++ Source/driver/Castro_F.H
@@ -130,9 +130,6 @@ extern "C"
      BL_FORT_FAB_ARG_3D(edge),
      const amrex::Real* dx, const int coord_type, const int idir);
 
-  void ca_normalize_species
-    (const int* lo, const int* hi, BL_FORT_FAB_ARG_3D(S_new));
-
   void ca_get_center(amrex::Real* center);
   void ca_set_center(const amrex::Real* center);
   void ca_find_center(amrex::Real* data, amrex::Real* center, const int* icen,
