### 2020-09-05 90a1715
remove some compound operators
