### 2020-09-06 14fde11
this compiles but NaNs
