+++ Exec/hydro_tests/gamma_law_bubble/Problem_Derive.cpp
@@ -4,9 +4,12 @@
 #include "Castro.H"
 #include "Castro_F.H"
 #include "prob_parameters.H"
+#include "extern_parameters.H"
+#include "gravity_params.H"
 
 using namespace amrex;
 
+using RealVector = amrex::Gpu::ManagedVector<amrex::Real>;
 
 void ca_derpi(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
               const FArrayBox& datfab, const Geometry& geomdata,
@@ -28,10 +31,15 @@ void ca_derpi(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
   int npts_1d = (2.0_rt*center[AMREX_SPACEDIM] + 1.e-8_rt) /
     dx[AMREX_SPACEDIM];
 
-  Array1D<Real, 0, npts_1d-1> pressure;
-  Array1D<Real, 0, npts_1d-1> density;
-  Array1D<Real, 0, npts_1d-1> temp;
-  Array1D<Real, 0, npts_1d-1> eint;
+  RealVector pressure_rv(npts_1d, 0);
+  RealVector density_rv(npts_1d, 0);
+  RealVector temp_rv(npts_1d, 0);
+  RealVector eint_rv(npts_1d, 0);
+
+  Real* const pressure = pressure_rv.dataPtr();
+  Real* const density = density_rv.dataPtr();
+  Real* const temp = temp_rv.dataPtr();
+  Real* const eint = eint_rv.dataPtr();
 
   pressure[0] = pres_base;
   density[0]  = dens_base;
@@ -42,7 +50,7 @@ void ca_derpi(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
 
   // compute the pressure scale height (for an isothermal, ideal-gas
   // atmosphere)
-  Real H = pres_base / dens_base / std::abs(Gravity::const_grav);
+  Real H = pres_base / dens_base / std::abs(gravity::const_grav);
 
   for (int j = 0; j < npts_1d; j++) {
 
@@ -51,18 +59,18 @@ void ca_derpi(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
 
      if (do_isentropic) {
        Real z = static_cast<Real>(j) * dx[AMREX_SPACEDIM];
-       density[j] = dens_base * 
-         (Gravity::const_grav * dens_base * (gamma_const - 1.0_rt) * z/
-          std::pow((gamma_const*pres_base) + 1.0_rt), 1.0_rt/(gamma_const - 1.0_rt));
+       density[j] = dens_base *
+         std::pow((gravity::const_grav * dens_base * (eos_gamma - 1.0_rt) * z/
+                   (eos_gamma*pres_base) + 1.0_rt), 1.0_rt/(eos_gamma - 1.0_rt));
      } else {
-       Real z = (static_cast<Real>(j)+HALF) * dx[AMREX_SPACEDIM];
+       Real z = (static_cast<Real>(j) + 0.5_rt) * dx[AMREX_SPACEDIM];
        density[j] = dens_base * std::exp(-z/H);
      }
 
      if (j > 0) {
         pressure[j] = pressure[j-1] -
           dx[AMREX_SPACEDIM] * 0.5_rt *
-          (density[j] + density[j-1]) * std::abs(Gravity::const_grav);
+          (density[j] + density[j-1]) * std::abs(gravity::const_grav);
      }
 
[…]
