### 2020-09-06 15f4af2
call the init from Fortran
