+++ Source/driver/Castro_setup.cpp
@@ -228,7 +228,7 @@ Castro::variableSetUp ()
   }
 
   // Initialize the Fortran Microphysics
-  microphysics_initialize(small_temp, small_dens);
+  ca_microphysics_init();
 
   // now initialize the C++ Microphysics
 #ifdef CXX_REACTIONS
