+++ Exec/gravity_tests/code_comp/Problem_Derive.cpp
@@ -7,85 +7,117 @@
 
 using namespace amrex;
 
-#ifdef __cplusplus
-extern "C"
+void ca_derextheating(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
+                      const FArrayBox& datfab, const Geometry& geomdata,
+                      Real /*time*/, const int* /*bcrec*/, int /*level*/)
 {
-#endif
 
-    // Note that in the following routines, we are NOT passing
-    // several variables to Fortran that would be unused.
+  const auto dx = geomdata.CellSizeArray();
+  const auto problo = geomdata.ProbLoArray();
 
-    // These routines are called in an MFIter loop, so we do not
-    // need to explicitly synchronize after GPU kernels.
+  GpuArray<Real, 3> center;
+  ca_get_center(center.begin());
 
-    void ca_derextheating(Real* der, const int* der_lo, const int* der_hi, const int* nvar,
-                                 const Real* data, const int* data_lo, const int* data_hi, const int* ncomp,
-                                 const int* lo, const int* hi,
-                                 const int* domain_lo, const int* domain_hi,
-                                 const Real* dx, const Real* xlo,
-                                 const Real* time, const Real* dt, const int* bcrec,
-                                 const int* level, const int* grid_no)
-    {
+  auto const dat = datfab.array();
+  auto const der = derfab.array();
 
-#pragma gpu
-        derextheating(der, AMREX_INT_ANYD(der_lo), AMREX_INT_ANYD(der_hi), *nvar,
-                      data, AMREX_INT_ANYD(data_lo), AMREX_INT_ANYD(data_hi), *ncomp,
-                      AMREX_INT_ANYD(lo), AMREX_INT_ANYD(hi),
-                      AMREX_INT_ANYD(domain_lo), AMREX_INT_ANYD(domain_hi),
-                      AMREX_REAL_ANYD(dx), *time);
+  amrex::ParallelFor(bx,
+  [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
+  {
+    Real height = problo[AMREX_SPACEDIM-1] + 
+      (static_cast<Real>(AMREX_D_PICK(i,j,k)) + 0.5_rt) * dx[AMREX_SPACEDIM-1] ;
+
+    if (height < 1.125e0_rt * 4.e8_rt) {
+
+      Real fheat = std::sin(8.e0_rt * M_PI * (height/ 4.e8_rt - 1.0_rt));
+      der(i,j,k,0) = heating_factor * fheat;
+
+    } else {
+      der(i,j,k,0) = 0.0_rt;
     }
 
-    void ca_deradinvariant(Real* der, const int* der_lo, const int* der_hi, const int* nvar,
-                                 const Real* data, const int* data_lo, const int* data_hi, const int* ncomp,
-                                 const int* lo, const int* hi,
-                                 const int* domain_lo, const int* domain_hi,
-                                 const Real* dx, const Real* xlo,
-                                 const Real* time, const Real* dt, const int* bcrec,
-                                 const int* level, const int* grid_no)
-    {
-
-#pragma gpu
-        deradinvariant(der, AMREX_INT_ANYD(der_lo), AMREX_INT_ANYD(der_hi), *nvar,
[…]
