### 2020-09-07 1a137c4
work on converting this to C++
