### 2020-09-07 8d270a7
add some accessors for radiation to C++
