+++ Exec/radiation_tests/RadSphere/problem_util.H
@@ -1,6 +1,8 @@
 #ifndef PROBLEM_UTIL_H
 #define PROBLEM_UTIL_H
 
+#include "fundamental_constants.H"
+
 // problem parameters
 //
 // we will specify the opacity as kappa = kapp_0 (nu_0 / nu)**3
@@ -20,10 +22,7 @@ constexpr Real kappa_0 = 1.e13_rt; // scattering opacity (1/cm)
 constexpr Real R_sphere = 0.02_rt; // sphere radius (cm)
 constexpr Real r_obs = 0.06_rt; // observer location (cm)
 
-// physical parameters
-constexpr Real T_0 = 50.0_rt * ev2erg / k_B; // ambient temp (K)
-constexpr Real T_sphere = 1500.0_rt * ev2erg / k_B; // sphere temp (K)
-
+constexpr Real T_sphere = 1500.0_rt * C::ev2erg / C::k_B; // sphere temp (K)
 
 // to guard against over/underflows, we define some 'safe' functions
 // (they come from Eric)
@@ -31,7 +30,7 @@ constexpr Real T_sphere = 1500.0_rt * ev2erg / k_B; // sphere temp (K)
 AMREX_GPU_HOST_DEVICE AMREX_INLINE
 Real
 sferfc(const Real arg1) {
-  return = std::erfc(amrex::max(-26.0_rt, amrex::min(26.0_rt, arg1)));
+  return std::erfc(amrex::max(-26.0_rt, amrex::min(26.0_rt, arg1)));
 }
 
 AMREX_GPU_HOST_DEVICE AMREX_INLINE
@@ -55,8 +54,8 @@ planck(const Real nu, const Real T) {
   // have units of erg / cm^3 / MeV.  As a result, we have one
   // less factor of hplanck.
 
-  Real B = (8.0_rt * M_PI * hplanck * nu * nu * nu  / (c_light * c_light * c_light)) /
-    (std::exp(hplanck * nu / (k_B * T)) - 1.0_rt);
+  Real B = (8.0_rt * M_PI * C::hplanck * nu * nu * nu  / (C::c_light * C::c_light * C::c_light)) /
+    (std::exp(C::hplanck * nu / (C::k_B * T)) - 1.0_rt);
   B = amrex::min(amrex::max(B, 1.e-50_rt), 1.e200_rt);
 
   return B;
@@ -75,11 +74,11 @@ F_radsphere(const Real r, const Real t, const Real nu) {
 
   Real kappa = kappa_0 * (nu_0 / nu) * (nu_0 / nu) * (nu_0 / nu);
 
-  Real erfc_term1 = sferfc(std::sqrt(3.0_rt * kappa / (4.0_rt * c_light * amrex::max(t, 1.e-50_rt))) *
-                           (r - R_sphere) - std::sqrt(c_light * t * kappa));
+  Real erfc_term1 = sferfc(std::sqrt(3.0_rt * kappa / (4.0_rt * C::c_light * amrex::max(t, 1.e-50_rt))) *
+                           (r - R_sphere) - std::sqrt(C::c_light * t * kappa));
 
-  Real erfc_term2 = sferfc(std::sqrt(3.0_rt * kappa / (4.0_rt * c_light * amrex::max(t, 1.e-50_rt))) *
-                           (r - R_sphere) + std::sqrt(c_light * t * kappa));
+  Real erfc_term2 = sferfc(std::sqrt(3.0_rt * kappa / (4.0_rt * C::c_light * amrex::max(t, 1.e-50_rt))) *
+                           (r - R_sphere) + std::sqrt(C::c_light * t * kappa));
 
[…]
