### 2020-09-07 c525043
remove unneeded center
