+++ Exec/radiation_tests/RadThermalWave/Problem_Derive.cpp
@@ -18,9 +18,6 @@ void dertexact(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
   const auto dx = geomdata.CellSizeArray();
   const auto problo = geomdata.ProbLoArray();
 
-  GpuArray<Real, 3> center;
-  ca_get_center(center.begin());
-
   auto const dat = datfab.array();
   auto const der = derfab.array();
 
