### 2020-09-07 39b7b28
remove some unused routines (#1196)
this also puts ca_normalize_species in sdc since that's the only place it is used.
