+++ Util/scripts/write_probdata.py
@@ -135,71 +135,80 @@ def get_next_line(fin):
     return line[:pos]
 
 
-def parse_param_file(param_file):
-    """read all the parameters in the parameter file and add valid
-    parameters to the params list.  This returns the parameter list.
+def parse_param_file(default_param_file, prob_param_file):
+    """read all the parameters in both the default parameter file and
+    the problem-specific parameter file and add valid parameters to the
+    params list.  This returns the parameter list.
     """
 
     err = 0
     params_list = []
 
-    try:
-        f = open(param_file, "r")
-    except IOError:
-        sys.exit("write_probdata.py: ERROR: file {} does not exist".format(param_file))
+    for param_file in [default_param_file, prob_param_file]:
 
-    line = get_next_line(f)
-
-    while line and not err:
-
-        # this splits the line into separate fields.  A field is a
-        # single word or a pair in parentheses like "(a, b)"
-        fields = re.findall(r'[\w\"\+\./\-]+|\([\w+\./\-]+\s*,\s*[\w\+\.\-]+\)', line)
-
-        if len(fields) < 3:
-            print("write_probdata.py: ERROR: missing one or more fields in parameter definition.")
-            err = 1
-            continue
-
-        current_param = Parameter()
-
-        current_param.var = fields[0]
-        current_param.dtype = fields[1]
-        current_param.value = fields[2]
-
-        # optional field: in namelist
-        try:
-            in_namelist = fields[3]
-            if in_namelist in ["y", "Y"]:
-                current_param.in_namelist = True
+        # The default parameter file has to exist, but the
+        # problem-specific parameter file is optional, so
+        # skip it if it's not present.
+        if os.path.isfile(param_file):
+            f = open(param_file, "r")
+        else:
+            if param_file == default_param_file:
+                sys.exit("write_probdata.py: ERROR: file {} does not exist".format(param_file))
             else:
-                current_param.in_namelist = False
-
-        except:
-            current_param.in_namelist = False
-
-
-        # optional field: size -- this can have the form (var, module)
-        # or just be a size
-        try:
-            size_info = fields[4]
-            if size_info[0] == "(":
-                size, module = re.findall(r"\w+", size_info)
-            else:
-                size = size_info
-                module = None
-
-        except:
-            size = 1
-            module = None
-
-        current_param.size = size
-        current_param.module = module
-
-        if not err == 1:
-            params_list.append(current_param)
+                continue
 
         line = get_next_line(f)
+
+        while line and not err:
+
+            # this splits the line into separate fields.  A field is a
+            # single word or a pair in parentheses like "(a, b)"
+            fields = re.findall(r'[\w\"\+\./\-]+|\([\w+\./\-]+\s*,\s*[\w\+\.\-]+\)', line)
+
+            if len(fields) < 3:
+                print("write_probdata.py: ERROR: missing one or more fields in parameter definition.")
+                err = 1
+                continue
+
+            current_param = Parameter()
+
+            current_param.var = fields[0]
+            current_param.dtype = fields[1]
+            current_param.value = fields[2]
+
+            # optional field: in namelist
+            try:
+                in_namelist = fields[3]
+                if in_namelist in ["y", "Y"]:
+                    current_param.in_namelist = True
+                else:
+                    current_param.in_namelist = False
+
+            except:
+                current_param.in_namelist = False
+
+
+            # optional field: size -- this can have the form (var, module)
+            # or just be a size
+            try:
+                size_info = fields[4]
+                if size_info[0] == "(":
+                    size, module = re.findall(r"\w+", size_info)
+                else:
+                    size = size_info
[…]
