+++ Source/driver/Castro_io.cpp
@@ -841,9 +841,7 @@ Castro::writeJobInfo (const std::string& dir, const Real io_time)
 
   runtime_pretty_print(jobinfo_file_name.dataPtr(), &jobinfo_file_length);
 
-#ifdef PROB_PARAMS
   prob_params_pretty_print(jobinfo_file_name.dataPtr(), &jobinfo_file_length);
-#endif
 
 }
 
