### 2020-09-07 d17c3be
put some common initial model stuff in its own file
