### 2020-09-08 #2513 @joshia5
@zingale, @eloisabentivegna following are my comments and questions:

1. Explicit instructions or commands for loading modules could be given in the installation
instructions, I found that python>=3.7 might not be preloaded on all systems

2. I could not find significant commits in the repo for author John bell.

3. I could run the test case but is there a specific version requirement on Paraview to visualize the output
*plt files?
What reader/format should I use while opening the files?

4. In the paper, comparisons are made with other packages but for the checklist requirement,
I cannot comment on whether these other packages are commonly used

5. While the README mentions how to contribute to the software & Report issues, i
could not find guidelines on how a third party can seek support.

6. As I do not have research expertise in this field, I cannot comment on completeness of references

7. license: it is mentioned in the commit comment of the license file that
castro has gone through paperwork for BSD license, does this mean
that the license is a BSD type ?

Thanks,
Aditya
