+++ Exec/radiation_tests/RadThermalWave/Problem_Derive.cpp
@@ -6,6 +6,7 @@
 #include "Castro_F.H"
 #include "fundamental_constants.H"
 #include "prob_parameters.H"
+#include <cmath>
 
 using namespace amrex;
 
@@ -71,9 +72,6 @@ void derterror(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
 
   const auto dx = geomdata.CellSizeArray();
   const auto problo = geomdata.ProbLoArray();
-
-  GpuArray<Real, 3> center;
-  ca_get_center(center.begin());
 
   auto const dat = datfab.array();
   auto const der = derfab.array();
