### 2020-09-08 a27a021
switch to using header file for fill_ext_src
