+++ Exec/gravity_tests/code_comp/Castro_ext_src.cpp
@@ -20,7 +20,7 @@ Castro::fill_ext_source (const Real time, const Real dt,
     // Next we correct the source terms in the conservative update to
     // time-center them.  Here we call ext_src(old, new), and then
     // in time_center_source_terms we subtract off 1/2 of the first S
-    // and add 1/2 of the new S. 
+    // and add 1/2 of the new S.
     //
     // Therefore, to get a properly time-centered source, generally
     // speaking, you always want to use the "new" state here.  That
@@ -30,26 +30,27 @@ Castro::fill_ext_source (const Real time, const Real dt,
     const auto dx = geom.CellSizeArray();
     const auto prob_lo = geom.ProbLoArray();
 
+    ext_src.setVal(0.0);
+
 #ifdef _OPENMP
 #pragma omp parallel
 #endif
     for (MFIter mfi(ext_src, TilingIfNotGPU()); mfi.isValid(); ++mfi)
     {
         const Box& bx = mfi.tilebox();
-
-        Array4<Real const> const snew = state_new.array(mfi);
         Array4<Real> const src = ext_src.array(mfi);
 
         amrex::ParallelFor(bx,
         [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
         {
-            auto x = prob_lo[0] + (Real(i) + 0.5_rt) * dx[0];
+            auto y = prob_lo[1] + (Real(j) + 0.5_rt) * dx[1];
 
-            src(i,j,k,UEINT) = snew(i,j,k,URHO) * 
-                (A1*std::exp(-(x - 0.5_rt)*(x - 0.5_rt)/(sigma1*sigma1)) + 
-                 A2*std::exp(-(x - 1.0_rt)*(x - 1.0_rt)/(sigma2*sigma2)));
+            if (y < 1.125_rt * 4.8e8_rt) {
+                auto fheat = std::sin(8.0_rt * M_PI * (y / 4.e8_rt - 1.0_rt));
 
-            src(i,j,k,UEDEN) = src(i,j,k,UEINT); 
+                src(i,j,k,UEDEN) = heating_factor * fheat;
+                src(i,j,k,UEINT) = src(i,j,k,UEDEN);
+            }
         });
     }
 }
