### 2020-09-08 90906da
ditch GPU_COMPATIBLE_PROBLEM requirement
