+++ Exec/gravity_tests/code_comp/Castro_ext_src.H
@@ -17,13 +17,18 @@ void do_ext_src(int i, int j, int k,
     const Real* problo = geomdata.ProbLo();
     const Real* dx = geomdata.CellSize();
 
-    auto x = problo[0] + (Real(i) + 0.5_rt) * dx[0];
+    for (int n = 0; n < NSRC; ++n) {
+        src(i,j,k,n) = 0.0_rt;
+    }
 
-    src(i,j,k,UEINT) = snew(i,j,k,URHO) * 
-        (A1*std::exp(-(x - 0.5_rt)*(x - 0.5_rt)/(sigma1*sigma1)) + 
-         A2*std::exp(-(x - 1.0_rt)*(x - 1.0_rt)/(sigma2*sigma2)));
+    auto y = problo[1] + (Real(j) + 0.5_rt) * dx[1];
 
-    src(i,j,k,UEDEN) = src(i,j,k,UEINT); 
+    if (y < 1.125_rt * 4.8e8_rt) {
+        auto fheat = std::sin(8.0_rt * M_PI * (y / 4.e8_rt - 1.0_rt));
+
+        src(i,j,k,UEDEN) = heating_factor * fheat;
+        src(i,j,k,UEINT) = src(i,j,k,UEDEN);
+    }
 }
 
 #endif
