+++ Source/sources/Castro_ext_src.H
@@ -1,6 +1,8 @@
 #ifndef CASTRO_EXT_SRC_H
 #define CASTRO_EXT_SRC_H
 
+#include "prob_parameters.H"
+
 using namespace amrex;
 
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
@@ -9,6 +11,7 @@ void do_ext_src(int i, int j, int k,
                 Array4<const Real> const& sold,
                 Array4<const Real> const& snew,
                 Array4<Real> const& src,
+                GpuArray<Real, 3> center,
                 const Real dt)
 {
     // const Real* problo = geomdata.ProbLo();
