### 2020-09-08 d2f64cb
dropped sold from do_ext_src
