+++ Exec/hydro_tests/heating_verification/Castro_ext_src.H
@@ -8,7 +8,6 @@ using namespace amrex;
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
 void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
-                Array4<const Real> const& sold,
                 Array4<const Real> const& snew,
                 Array4<Real> const& src,
                 GpuArray<Real, 3> center,
