+++ Exec/hydro_tests/test_convect/Castro_ext_src.H
@@ -11,7 +11,6 @@ constexpr Real L_x = 2.5e8_rt;
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
 void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
-                Array4<const Real> const& sold,
                 Array4<const Real> const& snew,
                 Array4<Real> const& src,
                 GpuArray<Real, 3> center,
