+++ Exec/science/planet/Castro_ext_src.H
@@ -12,7 +12,6 @@ constexpr Real beta = p_orb_vel * 2.0_rt / p_radius;
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
 void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
-                Array4<const Real> const& sold,
                 Array4<const Real> const& snew,
                 Array4<Real> const& src,
                 GpuArray<Real, 3> center,
