+++ Exec/science/wdmerger/Castro_ext_src.H
@@ -21,7 +21,6 @@ constexpr Real relaxation_damping_timescale = relaxation_damping_factor * dynami
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
 void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
-                Array4<const Real> const& sold,
                 Array4<const Real> const& snew,
                 Array4<Real> const& src,
                 GpuArray<Real, 3> center,
