+++ Exec/gravity_tests/hydrostatic_adjust/Castro_ext_src.H
@@ -13,7 +13,6 @@ constexpr Real t_stop = heating_time;
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
 void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
-                Array4<const Real> const& sold,
                 Array4<const Real> const& snew,
                 Array4<Real> const& src,
                 GpuArray<Real, 3> center,
@@ -33,8 +32,8 @@ void do_ext_src(int i, int j, int k,
 
         auto Hext = H_0 * std::exp(-((dist - r_0)*(dist - r_0))/(W_0*W_0));
 
-        src(i,j,k,UEINT) = sold(i,j,k,URHO) * Hext;
-        src(i,j,k,UEDEN) = sold(i,j,k,URHO) * Hext;
+        src(i,j,k,UEINT) = snew(i,j,k,URHO) * Hext;
+        src(i,j,k,UEDEN) = snew(i,j,k,URHO) * Hext;
 
     } else if (prob_type == 3) {
         const int ihe4 = network_spec_index("helium-4");
@@ -50,10 +49,10 @@ void do_ext_src(int i, int j, int k,
         auto dist = std::sqrt(x*x + y*y + z*z);
 
         auto Hext = H_0 * std::exp(-((dist - r_0)*(dist - r_0))/(W_0*W_0)) * 
-            sold(i,j,k,UFS-1+ihe4) / sold(i,j,k,URHO);
+            snew(i,j,k,UFS-1+ihe4) / snew(i,j,k,URHO);
 
-        src(i,j,k,UEINT) = sold(i,j,k,URHO) * Hext;
-        src(i,j,k,UEDEN) = sold(i,j,k,URHO) * Hext;
+        src(i,j,k,UEINT) = snew(i,j,k,URHO) * Hext;
+        src(i,j,k,UEDEN) = snew(i,j,k,URHO) * Hext;
     }
 }
 
