+++ Exec/hydro_tests/heating_verification/Castro_ext_src.H
@@ -8,7 +8,7 @@ using namespace amrex;
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
 void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
-                Array4<const Real> const& snew,
+                Array4<const Real> const& state,
                 Array4<Real> const& src,
                 GpuArray<Real, 3> center,
                 const Real dt)
@@ -18,7 +18,7 @@ void do_ext_src(int i, int j, int k,
 
     auto x = problo[0] + (Real(i) + 0.5_rt) * dx[0];
 
-    src(i,j,k,UEINT) = snew(i,j,k,URHO) * 
+    src(i,j,k,UEINT) = state(i,j,k,URHO) * 
         (A1*std::exp(-(x - 0.5_rt)*(x - 0.5_rt)/(sigma1*sigma1)) + 
          A2*std::exp(-(x - 1.0_rt)*(x - 1.0_rt)/(sigma2*sigma2)));
 

[…]
