+++ Source/radiation/Castro_radiation.cpp
@@ -1,11 +1,11 @@
 
-#include "Castro.H"
-#include "Castro_F.H"
+#include <Castro.H>
+#include <Castro_F.H>
 
 using std::string;
 
 #ifdef RADIATION
-#include "Radiation.H"
+#include <Radiation.H>
 
 using namespace amrex;
 
