+++ Source/driver/main.cpp
@@ -20,13 +20,13 @@
 #include <AMReX_AmrLevel.H>
 
 #ifdef HYPRE
-#include "_hypre_utilities.h"
+#include <_hypre_utilities.h>
 #endif
 
 #include <time.h>
 
-#include "Castro.H"
-#include "Castro_io.H"
+#include <Castro.H>
+#include <Castro_io.H>
 
 using namespace amrex;
 
