### 2020-09-09 2a22c00
add stage to action to check for double quotes in header includes
