### 2020-09-10 5dd40e6
switch include style to use <> instead of "" (#1205)
This addresses #331, and allows for us to override headers in problem directories.
This conversion was largely done with:
for i in git grep -l "#include \""; do sed -r -i -E "s/#include "(\w+.\w+)"/#include <\1>/g" $i; done
