+++ Source/driver/Derive.cpp
@@ -1,11 +1,11 @@
-#include "AMReX_REAL.H"
+#include <AMReX_REAL.H>
 
-#include "Derive.H"
-#include "Castro.H"
-#include "Castro_F.H"
+#include <Derive.H>
+#include <Castro.H>
+#include <Castro_F.H>
 
 #ifdef DIFFUSION
-#include "diffusion_util.H"
+#include <diffusion_util.H>
 #endif
 
 using namespace amrex;
