### 2020-09-10 40dff0f
more fixes to wdmerger
