+++ Exec/science/wdmerger/Castro_external.cpp
@@ -150,3 +150,4 @@ Castro::fill_ext_source (const Real time, const Real dt, const MultiFab& state_o
            AMREX_REAL_ANYD(prob_lo), AMREX_REAL_ANYD(dx), time, dt);
     }
 }
+
