+++ .github/workflows/c-linter.yml
@@ -31,7 +31,9 @@ jobs:
 
       - name: Check header includes
         run: |
-          if [[ $(grep -rIE --include="*.cpp" --include="*.H" "#\s?include\s+\"\w+\.\w+\"") ]]; then exit 1; fi
+          HEADER_INCLUDES=$(grep -rIE --include="*.cpp" --include="*.H" "#\s?include\s+\"\w+\.\w+\"")
+          echo $HEADER_INCLUDES
+          if [[ $HEADER_INCLUDES ]]; then exit 1; fi
 
       - name: Run cpp linter
         run: python3 external/cpp-linter-action/run_on_changed_files.py ${{ github.event.pull_request.base.sha }} ${{ github.event.pull_request.head.sha }} -header-filter=Castro -ignore-files="amrex|Microphysics" 
