### 2020-09-10 8944351
use damped momentum in wdmerger
