### 2020-09-11 8ac204a
pass GpuArrays by reference, pass time to do_ext_src, fix indexing
