+++ Exec/science/wdmerger/Castro_external.cpp
@@ -138,7 +138,7 @@ Castro::fill_ext_source (const Real time, const Real dt, const MultiFab& state_o
         amrex::ParallelFor(bx,
         [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
         {
-            do_ext_src(i, j, k, geomdata, snew, src, center, omega, time, dt);
+            do_ext_src(i, j, k, geomdata, snew, src, center, omega, dt, time);
         });
 
 #pragma gpu box(bx)
