+++ Source/sources/Castro_external.cpp
@@ -135,7 +135,7 @@ Castro::fill_ext_source (const Real time, const Real dt, const MultiFab& state_o
         amrex::ParallelFor(bx,
         [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
         {
-            do_ext_src(i, j, k, geomdata, snew, src, center, dt);
+            do_ext_src(i, j, k, geomdata, snew, src, center, dt, time);
         });
 
 #pragma gpu box(bx)
