### 2020-09-11 ad76031
fix y and z for DIM > 1 in hydrostatic_adjust
