+++ Source/sources/Castro_ext_src.H
@@ -13,11 +13,9 @@ void do_ext_src(int i, int j, int k,
                 GpuArray<Real, 3> const& center,
                 const Real dt, const Real time)
 {
-    // const Real* problo = geomdata.ProbLo();
-    // const Real* probhi = geomdata.ProbHi();
-    // const Real* dx     = geomdata.CellSize();
-    // const int*  domlo  = geomdata.domain.loVect();
-    // const int*  domhi  = geomdata.domain.hiVect();
+    // const auto problo = geomdata.ProbLoArray();
+    // const auto probhi = geomdata.ProbHiArray();
+    // const auto dx = geomdata.CellSizeArray();
 }
 
 #endif
