### 2020-09-11 eff4910
Convert ext_src to C++ (#1191)
This address #1187. fill_ext_source is moved from Castro_external.cpp into its own source file so that this can be overwritten for individual problems.
