+++ Exec/hydro_tests/test_convect/Castro_ext_src.H
@@ -1,0 +1,36 @@
+#ifndef CASTRO_EXT_SRC_H
+#define CASTRO_EXT_SRC_H
+
+#include <prob_parameters.H>
+
+using namespace amrex;
+
+constexpr Real y_layer = 1.25e8_rt;
+constexpr Real L_x = 2.5e8_rt;
+
+AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+void do_ext_src(int i, int j, int k,
+                GeometryData const& geomdata, 
+                Array4<const Real> const& state,
+                Array4<Real> const& src,
+                GpuArray<Real, 3> const& center,
+                const Real dt, const Real time)
+{
+    const Real* problo = geomdata.ProbLo();
+    const Real* dx = geomdata.CellSize();
+
+    auto x = (Real(i)+0.5e0_rt)*dx[0] + problo[0];
+    auto y = (Real(j)+0.5e0_rt)*dx[1] + problo[1];
+    auto ey = std::exp(-(y-y_layer)*(y-y_layer)/1.e14_rt);
+
+    auto H = ey * (1.e0_rt + 
+        0.00625_rt * std::sin( 2*M_PI*x/L_x) 
+        + 0.01875_rt * std::sin((6*M_PI*x/L_x) + M_PI/3.e0_rt) 
+        + 0.01250_rt * std::sin((8*M_PI*x/L_x) + M_PI/5.e0_rt));
+
+    // Source terms
+    src(i,j,k,UEDEN) = state(i,j,k,URHO) * H * 2.5e16_rt;
+    src(i,j,k,UEINT) = src(i,j,k,UEDEN);
+}
+
+#endif
