### 2020-09-11 40dd130
convert react_util.F90 to cpp
