+++ Source/sdc/Castro_sdc_util.H
@@ -1,0 +1,309 @@
+#ifndef CASTRO_SDC_UTIL_H
+#define CASTRO_SDC_UTIL_H
+
+#include <Castro.H>
+
+// error codes 
+constexpr int NEWTON_SUCCESS = 0;
+constexpr int SINGULAR_MATRIX = -1;
+constexpr int CONVERGENCE_FAILURE = -2;
+
+// solvers 
+constexpr int NEWTON_SOLVE = 1;
+constexpr int VODE_SOLVE = 2;
+constexpr int HYBRID_SOLVE = 3;
+
+
+AMREX_GPU_HOST_DEVICE AMREX_INLINE
+void
+normalize_species(const int i, const int j, const int k,
+                  Array4<const Real> const& u) {
+
+    GpuArray<Real, NumSpec> xn;
+
+    Real xn_sum = 0.0_rt;
+
+    for (int n = 0; n < NumSpec; ++n) {
+        xn[n] = u(i,j,k,UFS+n);
+        xn[n] = amrex::max::(small_x * u(i,j,k,URHO), amrex::min(u(i,j,k,URHO), xn[n]));
+        xn_sum += xn[n];
+    }
+
+    for (int n = 0; n < NumSpec; ++n) {
+        xn[n] = u(i,j,k,URHO) * (xn[n] / xn_sum);
+        u(i,j,k,UFS+n) = xn[n];
+    }
+}
+
+#ifdef REACTIONS
+
+AMREX_GPU_HOST_DEVICE AMREX_INLINE
+void
+sdc_update_o2(const int i, const int j, const int k,
+                 Array4<const Real> const& k_m, 
+                 Array4<Real> const& k_n, 
+                 Array4<const Real> const& A_m, 
+                 Array4<const Real> const& R_m_old,
+                 Array4<const Real> const& C, 
+                 const Real dt_m,
+                 const int sdc_iteration, const int m_start) {
+    // update k_m to k_n via advection -- this is a second-order accurate update
+
+    // Here, dt_m is the timestep between time-nodes m and m+1
+
+    burn_t burn_state;
+
+    GpuArray<Real, NVAR> U_old;
+    GpuArray<Real, NVAR> U_new;
+    GpuArray<Real, NVAR> R_full;
+    GpuArray<Real, NVAR> C_zone;
+
+    for (int n = 0; n < NVAR; ++n) {
+        U_old[n] = k_m(i,j,k,n);
+        C_zone[n] = C(i,j,k,n);
+    }
+
+    // only burn if we are within the temperature and density
+    // limits for burning
+    if (!okay_to_burn(U_old)) {
+        for (int n = 0; n < NVAR; ++n) {
+            R_full[n] = 0.0_rt;
+        }
+    } else {
+
+        // this is the full state -- this will be updated as we
+        // solve the nonlinear system.  We want to start with a
+        // good initial guess.  For later iterations, we should
+        // begin with the result from the previous iteration.  For
+        // the first iteration, let's try to extrapolate forward
+        // in time.
+        if (sdc_iteration == 0) {
+            for (int n = 0; n < NVAR; ++n) {
+                U_new[n] = U_old[n] + dt_m * A_m(i,j,k,n) + dt_m * R_m_old(i,j,k,n);
+            }
+        } else {
+            for (int n = 0; n < NVAR; ++n) {
+                U_new[n] = k_n(i,j,k,n);
+            }
+        }
+
+        sdc_solve(dt_m, U_old, U_new, C, sdc_iteration);
+
+        // we solved our system to some tolerance, but let's be sure we are conservative by
[…]
