### 2020-09-14 5878420
normalize_species and instantaneous_react compile and run without error
