+++ Source/reactions/Castro_react_util.H
@@ -2,6 +2,8 @@
 #define CASTRO_REACT_UTIL_H
 
 #include <Castro.H>
+#include <burn_type.H>
+#include <actual_rhs.H>
 
 
 AMREX_GPU_HOST_DEVICE AMREX_INLINE
@@ -36,7 +38,7 @@ single_zone_react_source(const int i, const int j, const int k,
              Array4<Real> const& R,
              burn_t burn_state) {
 
-    rhoInv = 1.0_rt / state(i,j,k,URHO);
+    auto rhoInv = 1.0_rt / state(i,j,k,URHO);
 
     burn_state.rho = state(i,j,k,URHO);
     burn_state.T = state(i,j,k,UTEMP);
@@ -59,139 +61,139 @@ single_zone_react_source(const int i, const int j, const int k,
     eos(eos_input_re, eos_state);
     eos_to_burn(eos_state, burn_state);
 
-    eos_get_small_temp(small_temp);
-    burn_state.T = amrex::min(max_temp, amrex::max(burn_state.T, small_temp));
+    eos_get_small_temp(&small_temp);
+    burn_state.T = amrex::min(MAX_TEMP, amrex::max(burn_state.T, small_temp));
 
     burn_state.self_heat = false;
 
-    GpuArray<Real, neqs> ydot;
+    Array1D<Real, 1, neqs> ydot;
 
     actual_rhs(burn_state, ydot);
 
     // store the instantaneous R
-    for (int n = 0; n < NVAR; ++n) {
+    for (int n = 0; n < NUM_STATE; ++n) {
         R(i,j,k,n) = 0.0_rt;
     }
 
     // species rates come back in terms of molar fractions
     for (int n = 0; n < NumSpec; ++n) {
-    R(i,j,k,UFS+n) = state(i,j,k,URHO) * aion[n] * ydot[n];
+        R(i,j,k,UFS+n) = state(i,j,k,URHO) * aion[n] * ydot(n);
     }
 
-    R(i,j,k,UEDEN) = state(i,j,k,URHO) * ydot[net_ienuc];
-    R(i,j,k,UEINT) = state(i,j,k,URHO) * ydot[net_ienuc];
+    R(i,j,k,UEDEN) = state(i,j,k,URHO) * ydot(net_ienuc);
+    R(i,j,k,UEINT) = state(i,j,k,URHO) * ydot(net_ienuc);
 }
 
-AMREX_GPU_HOST_DEVICE AMREX_INLINE
-void
-single_zone_jac(const int i, const int j, const int k,
-             Array4<const Real> const& state,
-             burn_t burn_state,
-             Array2D<Real, 0, NumSpec+1, 0, NumSpec+1> const& dRdw) {
+// AMREX_GPU_HOST_DEVICE AMREX_INLINE
+// void
+// single_zone_jac(const int i, const int j, const int k,
+//              Array4<const Real> const& state,
+//              burn_t burn_state,
+//              Array2D<Real, 0, NumSpec+1, 0, NumSpec+1> const& dRdw) {
 
-#ifdef SIMPLIFIED_SDC
-#ifndef AMREX_USE_GPU
-    amrex::Error("we shouldn't be here with the simplified SDC method (USE_SIMPLIFIED_SDC=TRUE)");
-#endif
-#else
+// #ifdef SIMPLIFIED_SDC
+// #ifndef AMREX_USE_GPU
+//     amrex::Error("we shouldn't be here with the simplified SDC method (USE_SIMPLIFIED_SDC=TRUE)");
+// #endif
+// #else
 
-    GpuArray<Real, neqs> ydot;
[…]
