+++ Source/sdc/Castro_sdc_util.H
@@ -2,6 +2,7 @@
 #define CASTRO_SDC_UTIL_H
 
 #include <Castro.H>
+#include <Castro_react_util.H>
 
 // error codes 
 constexpr int NEWTON_SUCCESS = 0;
@@ -16,8 +17,8 @@ constexpr int HYBRID_SOLVE = 3;
 
 AMREX_GPU_HOST_DEVICE AMREX_INLINE
 void
-normalize_species(const int i, const int j, const int k,
-                  Array4<const Real> const& u) {
+normalize_species_sdc(const int i, const int j, const int k,
+                  Array4<Real> const& u) {
 
     GpuArray<Real, NumSpec> xn;
 
@@ -25,7 +26,7 @@ normalize_species(const int i, const int j, const int k,
 
     for (int n = 0; n < NumSpec; ++n) {
         xn[n] = u(i,j,k,UFS+n);
-        xn[n] = amrex::max::(small_x * u(i,j,k,URHO), amrex::min(u(i,j,k,URHO), xn[n]));
+        xn[n] = amrex::max(small_x * u(i,j,k,URHO), amrex::min(u(i,j,k,URHO), xn[n]));
         xn_sum += xn[n];
     }
 
@@ -35,275 +36,281 @@ normalize_species(const int i, const int j, const int k,
     }
 }
 
-#ifdef REACTIONS
+// #ifdef REACTIONS
 
-AMREX_GPU_HOST_DEVICE AMREX_INLINE
-void
-sdc_update_o2(const int i, const int j, const int k,
-                 Array4<const Real> const& k_m, 
-                 Array4<Real> const& k_n, 
-                 Array4<const Real> const& A_m, 
-                 Array4<const Real> const& R_m_old,
-                 Array4<const Real> const& C, 
-                 const Real dt_m,
-                 const int sdc_iteration, const int m_start) {
-    // update k_m to k_n via advection -- this is a second-order accurate update
+// AMREX_GPU_HOST_DEVICE AMREX_INLINE
+// void
+// sdc_update_o2(const int i, const int j, const int k,
+//                  Array4<const Real> const& k_m, 
+//                  Array4<Real> const& k_n, 
+//                  Array4<const Real> const& A_m, 
+//                  Array4<const Real> const& R_m_old,
+//                  Array4<const Real> const& C, 
+//                  const Real dt_m,
+//                  const int sdc_iteration, const int m_start) {
+//     // update k_m to k_n via advection -- this is a second-order accurate update
 
-    // Here, dt_m is the timestep between time-nodes m and m+1
+//     // Here, dt_m is the timestep between time-nodes m and m+1
 
-    burn_t burn_state;
+//     burn_t burn_state;
 
-    GpuArray<Real, NVAR> U_old;
-    GpuArray<Real, NVAR> U_new;
-    GpuArray<Real, NVAR> R_full;
-    GpuArray<Real, NVAR> C_zone;
+//     GpuArray<Real, NUM_STATE> U_old;
+//     GpuArray<Real, NUM_STATE> U_new;
+//     GpuArray<Real, NUM_STATE> R_full;
+//     GpuArray<Real, NUM_STATE> C_zone;
 
-    for (int n = 0; n < NVAR; ++n) {
-        U_old[n] = k_m(i,j,k,n);
-        C_zone[n] = C(i,j,k,n);
-    }
+//     for (int n = 0; n < NUM_STATE; ++n) {
+//         U_old[n] = k_m(i,j,k,n);
+//         C_zone[n] = C(i,j,k,n);
+//     }
 
-    // only burn if we are within the temperature and density
-    // limits for burning
-    if (!okay_to_burn(U_old)) {
[…]
