+++ Source/hydro/Castro_ctu_hydro.cpp
@@ -2,7 +2,6 @@
 #include <Castro_util.H>
 #include <Castro_F.H>
 #include <Castro_hydro.H>
-#include <Castro_hydro_F.H>
 
 #ifdef RADIATION
 #include <Radiation.H>
