### 2020-09-15 dcbc635
fix GPU compilation (#1211)
