+++ Exec/hydro_tests/toy_convect/Castro_ext_src.H
@@ -5,7 +5,7 @@
 
 using namespace amrex;
 
-AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
 void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
                 Array4<const Real> const& state,
