### 2020-09-15 #2513 @zingale
Thanks for the comments.  I will answer them each inline here.

1. Can you point out where we suggest python 3.7?  It looks like we
   require 3.6.  If you have suggestions for what types of machines
   you think there should be instructions for, please share.  There
   are so many variations that it would be hard to cover them all.

   Python version 3.5 has reached "end of life":
   https://www.python.org/downloads/release/python-3510/ so we believe
   requiring 3.6 or later is reasonable.

   For working at supercomputing centers, we do provide instructions,
   linked to from this page of the docs (at the bottom):
   https://amrex-astro.github.io/Castro/docs/mpi_plus_x.html
