### 2020-09-15 #2513 @zingale
2.  Not all Castro developers write code.  John contributed the initial
   version of the hydro solver (before Castro was in version control)
   and he works on the mathematical aspects of the algorithm
   development (most recently the SDC stuff).
