### 2020-09-15 #2513 @zingale
3. We don't have experience with Paraview, so we will remove it's mention
   from the docs.  All of the developers use yt.
