### 2020-09-15 #2513 @zingale
5. The last section of the README has a link to the castro-help mailing list.
   Further, the mailing list is featured under the "get help" button on the
   main website: https://amrex-astro.github.io/Castro/
