### 2020-09-15 #2513 @zingale
7. The license is OSI approved: https://opensource.org/BSD-3-Clause-LBNL
