### 2020-09-15 ab0eba6
remove reference to paraview + add workflow docs (#1212)
