+++ Source/sdc/Castro_sdc_util.H
@@ -3,6 +3,7 @@
 
 #include <Castro.H>
 #include <Castro_react_util.H>
+#include <Castro_sdc_vode.H>
 
 // error codes 
 constexpr int NEWTON_SUCCESS = 0;
@@ -38,71 +39,569 @@ normalize_species_sdc(const int i, const int j, const int k,
 
 // #ifdef REACTIONS
 
-// AMREX_GPU_HOST_DEVICE AMREX_INLINE
-// void
-// sdc_update_o2(const int i, const int j, const int k,
-//                  Array4<const Real> const& k_m, 
-//                  Array4<Real> const& k_n, 
-//                  Array4<const Real> const& A_m, 
-//                  Array4<const Real> const& R_m_old,
-//                  Array4<const Real> const& C, 
-//                  const Real dt_m,
-//                  const int sdc_iteration, const int m_start) {
-//     // update k_m to k_n via advection -- this is a second-order accurate update
+AMREX_GPU_HOST_DEVICE AMREX_INLINE
+void
+sdc_solve(const int i, const int j, const int k,
+          Array4<const Real> const& U_old, 
+          Array4<Real> const& U_new, 
+          Array4<const Real> const& C, 
+          const Real dt_m,
+          const int sdc_iteration) {
 
-//     // Here, dt_m is the timestep between time-nodes m and m+1
+    int ierr;
 
-//     burn_t burn_state;
+    // for debugging
+    GpuArray<Real, NUM_STATE> U_orig;
 
-//     GpuArray<Real, NUM_STATE> U_old;
-//     GpuArray<Real, NUM_STATE> U_new;
-//     GpuArray<Real, NUM_STATE> R_full;
-//     GpuArray<Real, NUM_STATE> C_zone;
+    for (int n = 0; n < NUM_STATE; ++n) {
+        U_orig[n] = U_old(i,j,k,n);
+    }
 
-//     for (int n = 0; n < NUM_STATE; ++n) {
-//         U_old[n] = k_m(i,j,k,n);
-//         C_zone[n] = C(i,j,k,n);
-//     }
+    if (sdc_solver == NEWTON_SOLVE) {
+        // we are going to assume we already have a good guess for the
+        // solving in U_new and just pass the solve onto the main Newton
+        // solve
+        auto err_out = sdc_newton_subdivide(dt_m, U_old, U_new, C, sdc_iteration, &ierr);
 
-//     // only burn if we are within the temperature and density
-//     // limits for burning
-//     if (!okay_to_burn(U_old)) {
-//         for (int n = 0; n < NUM_STATE; ++n) {
-//             R_full[n] = 0.0_rt;
-//         }
-//     } else {
+        // failing?
+        if (ierr != NEWTON_SUCCESS) {
+            AllPrint() << "Newton convergence failure" << std::endl;
+            AllPrint() << "convergence failure, error = " << err_out << std::endl;
+            AllPrint() << "density: " << U_old(i,j,k,URHO) << std::endl;
+            AllPrint() << "(old) temperature: " << U_old(i,j,k,UTEMP) << std::endl;
+            AllPrint() << "mass fractions:" 
+            for (int n = 0; n < NumSpec; ++n) {
+                AllPrint() << " " << U_old(i,j,k,UFS+n)/U_old(i,j,k,URHO);
+            }
+            AllPrint() << std::endl;
+            Abort("Newton subcycling failed in sdc_solve");
+        }
 
-//         // this is the full state -- this will be updated as we
[…]
