### 2020-09-16 faffb0e
some docs updates CITATIONS was moved to markdown and the papers updated
the list of core developers was removed from the Preface and linked instead to the webpage.
This way they don't get out of sync.
