+++ CITATION.md
@@ -1,0 +1,198 @@
+# Citing Castro
+
+There are a number of Castro papers that describe parts of the
+algorithm.  We ask that you cite all of the appropriate papers
+describing the capabilities you used.
+
+## General code use
+
+If you use Castro, we appreciate you citing the main code paper:
+
+```
+    @ARTICLE{2010ApJ...715.1221A,
+       author = {{Almgren}, A.~S. and {Beckner}, V.~E. and {Bell},
+		      J.~B. and {Day}, M.~S. and {Howell}, L.~H. and
+		      {Joggerst}, C.~C. and {Lijewski}, M.~J. and
+		      {Nonaka}, A. and {Singer}, M. and {Zingale}, M.},
+	title = "{CASTRO: A New Compressible Astrophysical
+		      Solver. I. Hydrodynamics and Self-gravity}",
+      journal = {\apj},
+    archivePrefix = "arXiv",
+       eprint = {1005.0114},
+     primaryClass = "astro-ph.IM",
+     keywords = {equation of state, gravitation, hydrodynamics, methods:
+		      numerical, nuclear reactions, nucleosynthesis,
+		      abundances},
+	 year = 2010,
+	month = jun,
+       volume = 715,
+	pages = {1221-1238},
+	  doi = {10.1088/0004-637X/715/2/1221},
+       adsurl = {http://adsabs.harvard.edu/abs/2010ApJ...715.1221A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
+```
+
+You should also cite the zenodo DOI for the code release.  A bibtex
+entry for the latest release can be found here:
+
+https://doi.org/10.5281/zenodo.2301848
+
+## Radiation hydrodynamics
+
+If you use the radiation hydrodynamics capabilities, please additionally
+cite the following:
+
+```
+    @ARTICLE{2011ApJS..196...20Z,
+       author = {{Zhang}, W. and {Howell}, L. and {Almgren}, A. and
+		      {Burrows}, A. and {Bell}, J.},
+	title = "{CASTRO: A New Compressible Astrophysical
+		      Solver. II. Gray Radiation Hydrodynamics}",
+      journal = {\apjs},
+    archivePrefix = "arXiv",
+       eprint = {1105.2466},
+     primaryClass = "astro-ph.IM",
+     keywords = {diffusion, hydrodynamics, methods: numerical, radiative
+		      transfer},
+	 year = 2011,
+	month = oct,
+       volume = 196,
+	  eid = {20},
+	pages = {20},
+	  doi = {10.1088/0067-0049/196/2/20},
+       adsurl = {http://adsabs.harvard.edu/abs/2011ApJS..196...20Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
+```
+
+```
+    @ARTICLE{2013ApJS..204....7Z,
+       author = {{Zhang}, W. and {Howell}, L. and {Almgren}, A. and
+		      {Burrows}, A. and {Dolence}, J. and {Bell}, J.},
+	title = "{CASTRO: A New Compressible Astrophysical
+		      Solver. III. Multigroup Radiation Hydrodynamics}",
+      journal = {\apjs},
+    archivePrefix = "arXiv",
+       eprint = {1207.3845},
+     primaryClass = "astro-ph.IM",
+     keywords = {diffusion, hydrodynamics, methods: numerical, radiative
+		      transfer },
+	 year = 2013,
+	month = jan,
+       volume = 204,
+	  eid = {7},
+	pages = {7},
+	  doi = {10.1088/0067-0049/204/1/7},
+       adsurl = {http://adsabs.harvard.edu/abs/2013ApJS..204....7Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
+```
+
[…]
