### 2020-09-17 #2513 @eloisabentivegna
@zingale, thanks for the replies. There is nothing for you to do regarding points 4. and 6. I will now wait for @joshia5 to comment back and share his final recommendation.
