### 2020-09-18 41cbb75
the size of the reactions array was incorrectly dimensioned
