### 2020-09-18 92c8a8d
don't use dtnuc_e if the region is in NSE
