### 2020-09-18 b2a9bba
convert estdt_burning to C++
