+++ CHANGES.md
@@ -1,5 +1,13 @@
 # 20.10
 
+   * Automatic problem parameter configuration is now available to every
+     problem by placing a _prob_params file in your problem directory.
+     Examples can be found in most of the problems in Castro/Exec, and you
+     can look at the "Setting Up Your Own Problem" section of the documentation
+     for more information. This functionality is optional, however note that
+     a file named probdata.F90 is now automatically generated, so if you have
+     a legacy file with that name it should be renamed. (#1210)
+
    * The external heat source term routines have been ported to C++
      (#1191).  Any problem using an external heat source should look
      convert the code over to C++.
