+++ Docs/source/build_system.rst
@@ -19,7 +19,7 @@ space for variables that are not used.
 General Build Parameters
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
-.. index:: USE_ALL_CASTRO, USE_AMR_CORE, USE_SYSTEM_BLAS, USE_HYPRE, USE_PROB_PARAMS
+.. index:: USE_ALL_CASTRO, USE_AMR_CORE, USE_SYSTEM_BLAS, USE_HYPRE
 
 These Parameters affect the build (parallelism, performance, etc.)
 Most of these are parameters from AMReX.
@@ -46,9 +46,6 @@ Most of these are parameters from AMReX.
   * ``USE_HYPRE``: compile in the Hypre library.  This will be automatically enabled
     for radiation.  You need to specify the path to the Hypre library via either
     ``HYPRE_DIR`` or ``HYPRE_OMP_DIR``.
-
-  * ``USE_PROB_PARAMS``: generate the ``probdata_module`` at runtime by parsing
-    the problem's ``_prob_params`` file.
 
 Parallelization and GPUs
 ^^^^^^^^^^^^^^^^^^^^^^^^
@@ -273,14 +270,14 @@ This is the current build system process.
 
 * Problem-specific runtime parameters are parsed by ``write_probdata.py``
 
-  * If ``USE_PROB_PARAMS = TRUE``, then the ``_prob_param`` file in
-    the problem directory is parsed and used to define the Fortran
-    ``&fortin`` namelist that controls the runtime parameters for
-    problem initialization.
+  * If the problem directory defines a ``_prob_params`` then it is parsed
+    and used to define the Fortran ``&fortin`` namelist that controls the
+    runtime parameters for problem initialization. Either way, the namelist
+    will include all variables in ``Castro/Source/problems/_default_prob_params``.
 
   * The script ``Castro/Util/scripts/write_probdata.py`` is used
 
-  * The hook for this is in ``Make.Castro`` in the ``prob_params_auto.F90`` rule.
+  * The hook for this is in ``Make.auto_source`` in the ``prob_params_auto.F90`` rule.
 
   * The ``prob_params_auto.F90`` file is output into ``tmp_build_dir/castro_sources/``.
 
