### 2020-09-22 1dd5cc8
Convert center to auto prob params
