+++ Source/driver/prob_params_nd.F90
@@ -14,7 +14,7 @@ module prob_params_module
 
   ! geometry information
   integer,  allocatable, save :: coord_type
-  real(rt), allocatable :: center(:), problo(:), probhi(:)
+  real(rt), allocatable :: problo(:), probhi(:)
 
   ! dimension information
   integer, save, allocatable :: dim
@@ -53,7 +53,7 @@ module prob_params_module
   attributes(managed) :: dim
   attributes(managed) :: dg
   attributes(managed) :: coord_type
-  attributes(managed) :: center, problo, probhi
+  attributes(managed) :: problo, probhi
   attributes(managed) :: domlo_level, domhi_level, dx_level
   attributes(managed) :: ref_ratio, n_error_buf, blocking_factor
   attributes(managed) :: mom_flux_has_p
@@ -64,7 +64,7 @@ module prob_params_module
   !$acc declare create(dim)
   !$acc declare create(dg)
   !$acc declare create(coord_type)
-  !$acc declare create(center, problo, probhi)
+  !$acc declare create(problo, probhi)
   !$acc declare create(domlo_level, domhi_level, dx_level)
   !$acc declare create(ref_ratio, n_error_buf, blocking_factor)
   !$acc declare create(mom_flux_has_p)
