+++ Source/driver/Castro_setup.cpp
@@ -324,14 +324,9 @@ Castro::variableSetUp ()
 
   const int coord_type = dgeom.Coord();
 
-  // Get the center variable from the inputs and pass it directly to Fortran.
-  Vector<Real> center(BL_SPACEDIM, 0.0);
-  ParmParse ppc("castro");
-  ppc.queryarr("center",center,0,BL_SPACEDIM);
-
   ca_set_problem_params(dm,phys_bc.lo(),phys_bc.hi(),
                         Interior,Inflow,Outflow,Symmetry,SlipWall,NoSlipWall,coord_type,
-                        dgeom.ProbLo(),dgeom.ProbHi(),center.dataPtr());
+                        dgeom.ProbLo(),dgeom.ProbHi());
 
   // Read in the parameters for the tagging criteria
   // and store them in the Fortran module.
