+++ Source/driver/Castro_util_nd.F90
@@ -108,39 +108,6 @@ contains
   end subroutine ca_clamp_temp
 
 
-  subroutine ca_get_center(center_out) bind(C, name="ca_get_center")
-    ! Get the current center of the problem.  This may not be the
-    ! center of the domain, due to any problem symmetries.
-
-    use prob_params_module, only: center
-    use amrex_fort_module, only: rt => amrex_real
-
-    implicit none
-
-    real(rt), intent(inout) :: center_out(3)
-
-    center_out = center
-
-  end subroutine ca_get_center
-
-
-  subroutine ca_set_center(center_in) bind(C, name="ca_set_center")
-    !
-    ! .. note::
-    !    Binds to C function ``ca_set_center``
-
-    use prob_params_module, only: center
-    use amrex_fort_module, only: rt => amrex_real
-
-    implicit none
-
-    real(rt), intent(in) :: center_in(3)
-
-    center = center_in
-
-  end subroutine ca_set_center
-
-
   subroutine ca_find_center(data,new_center,icen,dx,problo) &
        bind(C, name="ca_find_center")
     !
@@ -223,7 +190,8 @@ contains
     !    Binds to C function ``ca_compute_avgstate``
 
     use meth_params_module, only: URHO, UMX, UMY, UMZ
-    use prob_params_module, only: center, dim
+    use prob_params_module, only: dim
+    use probdata_module, only: center
     use amrex_constants_module, only: HALF
 #ifndef AMREX_USE_CUDA
     use castro_error_module, only: castro_error
