+++ Exec/science/wdmerger/Castro_ext_src.H
@@ -18,7 +18,6 @@ void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
                 Array4<const Real> const& state,
                 Array4<Real> const& src,
-                GpuArray<Real, 3> const& center,
                 GpuArray<Real, 3> const& omega,
                 const Real dt, const Real time)
 {

[…]
