### 2020-09-22 cc05903
Fix initialization order
