+++ Source/driver/Castro.H
@@ -10,6 +10,7 @@
 // runtime parameters
 #include <castro_params.H>
 #include <prob_parameters.H>
+#include <prob_parameters_F.H>
 
 using namespace castro;
 
