### 2020-09-22 df451d6
Add a prob_params redirect for center
