### 2020-09-22 5ee17ff
Add riemann.h constants to a namespace
