+++ Exec/science/wdmerger/Prob.cpp
@@ -75,25 +75,39 @@ Castro::wd_update (Real time, Real dt)
     BL_ASSERT(level == 0 || (!parent->subCycle() && level == parent->finestLevel()));
 
     // Get the current stellar data
-    get_star_data(com_p, com_s, vel_p, vel_s, &mass_p, &mass_s, &t_ff_p, &t_ff_s);
+    get_f90_com_P(com_P);
+    get_f90_com_S(com_S);
+    get_f90_vel_P(vel_P);
+    get_f90_vel_S(vel_S);
+    get_f90_mass_P(&mass_P);
+    get_f90_mass_S(&mass_S);
+    get_f90_t_ff_P(&t_ff_P);
+    get_f90_t_ff_S(&t_ff_S);
 
     // Update the problem center using the system bulk velocity
     update_center(&time);
 
     for ( int i = 0; i < 3; i++ ) {
-      com_p[i] += vel_p[i] * dt;
-      com_s[i] += vel_s[i] * dt;
+      com_P[i] += vel_P[i] * dt;
+      com_S[i] += vel_S[i] * dt;
     }
 
     // Now send this first estimate of the COM to Fortran, and then re-calculate
     // a more accurate result using it as a starting point.
 
-    set_star_data(com_p, com_s, vel_p, vel_s, &mass_p, &mass_s, &t_ff_p, &t_ff_s);
+    set_f90_com_P(com_P);
+    set_f90_com_S(com_S);
+    set_f90_vel_P(vel_P);
+    set_f90_vel_S(vel_S);
+    get_f90_mass_P(&mass_P);
+    get_f90_mass_S(&mass_S);
+    get_f90_t_ff_P(&t_ff_P);
+    get_f90_t_ff_S(&t_ff_S);
 
     // Save relevant current data.
 
-    Real old_mass_p = mass_p;
-    Real old_mass_s = mass_s;
+    Real old_mass_P = mass_P;
+    Real old_mass_S = mass_S;
 
     ReduceOps<ReduceOpSum, ReduceOpSum, ReduceOpSum, ReduceOpSum, ReduceOpSum, ReduceOpSum,
               ReduceOpSum, ReduceOpSum, ReduceOpSum, ReduceOpSum, ReduceOpSum, ReduceOpSum,
@@ -254,28 +268,28 @@ Castro::wd_update (Real time, Real dt)
 
               }
 
-              Real com_p_x = dmSymmetric * rSymmetric[0] * primary_factor;
-              Real com_p_y = dmSymmetric * rSymmetric[1] * primary_factor;
-              Real com_p_z = dmSymmetric * rSymmetric[2] * primary_factor;
+              Real com_P_x = dmSymmetric * rSymmetric[0] * primary_factor;
+              Real com_P_y = dmSymmetric * rSymmetric[1] * primary_factor;
+              Real com_P_z = dmSymmetric * rSymmetric[2] * primary_factor;
 
-              Real com_s_x = dmSymmetric * rSymmetric[0] * secondary_factor;
-              Real com_s_y = dmSymmetric * rSymmetric[1] * secondary_factor;
-              Real com_s_z = dmSymmetric * rSymmetric[2] * secondary_factor;
+              Real com_S_x = dmSymmetric * rSymmetric[0] * secondary_factor;
+              Real com_S_y = dmSymmetric * rSymmetric[1] * secondary_factor;
+              Real com_S_z = dmSymmetric * rSymmetric[2] * secondary_factor;
 
-              Real vel_p_x = momSymmetric[0] * vol(i,j,k) * primary_factor;
[…]
