### 2020-09-22 c78eeb7
Problem parameters not in namelist also have a C++ version (#1224)
Previously you could add parameters to _prob_params that were not in the Fortran namelist. No C++ version was made of these parameters. However, there is a need for even the ones not in the namelist to have a C++ version, since we may set those variables in probinit and want them later in C++. This implements that change, with the exception of parameters whose size depends on a Fortran module parameter (since we have no direct analogue of that in C++). As a demonstration, many of the wdmerger C++ problem parameters (which were 
[…]
