+++ Exec/science/wdmerger/sum_integrated_quantities.cpp
@@ -65,7 +65,7 @@ Castro::sum_integrated_quantities ()
 
     // Mass transfer rate
 
-    Real mdot = 0.5 * (std::abs(mdot_p) + std::abs(mdot_s));
+    Real mdot = 0.5 * (std::abs(mdot_P) + std::abs(mdot_S));
 
     // Center of mass of the system.
 
@@ -82,17 +82,17 @@ Castro::sum_integrated_quantities ()
 
     // Stellar centers of mass and velocities.
 
-    Real com_p_mag = 0.0;
-    Real com_s_mag = 0.0;
+    Real com_P_mag = 0.0;
+    Real com_S_mag = 0.0;
 
-    Real vel_p_mag = 0.0;
-    Real vel_s_mag = 0.0;
+    Real vel_P_mag = 0.0;
+    Real vel_S_mag = 0.0;
 
-    Real vel_p_rad = 0.0;
-    Real vel_s_rad = 0.0;
+    Real vel_P_rad = 0.0;
+    Real vel_S_rad = 0.0;
 
-    Real vel_p_phi = 0.0;
-    Real vel_s_phi = 0.0;
+    Real vel_P_phi = 0.0;
+    Real vel_S_phi = 0.0;
 
     // Gravitational wave amplitudes.
     
@@ -281,39 +281,39 @@ Castro::sum_integrated_quantities ()
 
     }
 
-    com_p_mag += std::pow( std::pow(com_p[0],2) + std::pow(com_p[1],2) + std::pow(com_p[2],2), 0.5 );
-    com_s_mag += std::pow( std::pow(com_s[0],2) + std::pow(com_s[1],2) + std::pow(com_s[2],2), 0.5 );
-    vel_p_mag += std::pow( std::pow(vel_p[0],2) + std::pow(vel_p[1],2) + std::pow(vel_p[2],2), 0.5 );
-    vel_s_mag += std::pow( std::pow(vel_s[0],2) + std::pow(vel_s[1],2) + std::pow(vel_s[2],2), 0.5 );
+    com_P_mag += std::pow( std::pow(com_P[0],2) + std::pow(com_P[1],2) + std::pow(com_P[2],2), 0.5 );
+    com_S_mag += std::pow( std::pow(com_S[0],2) + std::pow(com_S[1],2) + std::pow(com_S[2],2), 0.5 );
+    vel_P_mag += std::pow( std::pow(vel_P[0],2) + std::pow(vel_P[1],2) + std::pow(vel_P[2],2), 0.5 );
+    vel_S_mag += std::pow( std::pow(vel_S[0],2) + std::pow(vel_S[1],2) + std::pow(vel_S[2],2), 0.5 );
 
 #if (BL_SPACEDIM == 3)
-    if (mass_p > 0.0) {
-      vel_p_rad = (com_p[axis_1 - 1] / com_p_mag) * vel_p[axis_1 - 1] + (com_p[axis_2 - 1] / com_p_mag) * vel_p[axis_2 - 1];
-      vel_p_phi = (com_p[axis_1 - 1] / com_p_mag) * vel_p[axis_2 - 1] - (com_p[axis_2 - 1] / com_p_mag) * vel_p[axis_1 - 1];
+    if (mass_P > 0.0) {
+      vel_P_rad = (com_P[axis_1 - 1] / com_P_mag) * vel_P[axis_1 - 1] + (com_P[axis_2 - 1] / com_P_mag) * vel_P[axis_2 - 1];
[…]
