+++ Util/scripts/write_probdata.py
@@ -24,10 +24,8 @@ This script takes a template file and replaces keywords in it
 initialize the parameters, setup a namelist, set the defaults, etc.
 
 Note: there are two types of parameters here, the ones that are in the
-namelist are true runtime parameters.  For those we also provide the
-C++ headers and get routines to make these runtime parametrs available
-in C++.  The non-namelist parameters should be avoided if at all
-possible.
+namelist are true runtime parameters.  The non-namelist parameters
+should be avoided if at all possible.
 
 """
 
@@ -117,6 +115,9 @@ class Parameter:
                 return False
             else:
                 return True
+
+    def has_module(self):
+        return self.module != None
 
     def __str__(self):
         return "{} : {}, {} elements".format(self.var, self.dtype, self.size)
@@ -376,11 +377,12 @@ def write_probin(probin_template, default_prob_param_file, prob_param_file, out_
 
             elif keyword == "cxx_gets":
                 # this writes out the Fortran functions that can be called from C++
-                # to get the value of the parameters.  Note: we only do this for namelist
-                # parameters
+                # to get the value of the parameters.
 
                 for p in params:
-                    if not p.in_namelist:
+                    # We don't currently support the case where a module
+                    # is required since we have no C++ equivalent for the module.
+                    if p.is_array() and p.has_module():
                         continue
                     if p.dtype == "logical":
                         continue
@@ -406,8 +408,12 @@ def write_probin(probin_template, default_prob_param_file, prob_param_file, out_
                     else:
                         fout.write("{}subroutine get_f90_{}({}_in) bind(C, name=\"get_f90_{}\")\n".format(
                             indent, p.var, p.var, p.var))
-                        fout.write("{}   {}, intent(inout) :: {}_in\n".format(
-                            indent, p.get_f90_decl(), p.var))
+                        if p.is_array():
+                            fout.write("{}   {}, intent(inout) :: {}_in({})\n".format(
+                                indent, p.get_f90_decl(), p.var, p.size))
+                        else:
+                            fout.write("{}   {}, intent(inout) :: {}_in\n".format(
+                                indent, p.get_f90_decl(), p.var))
                         fout.write("{}   {}_in = {}\n".format(
                             indent, p.var, p.var))
                         fout.write("{}end subroutine get_f90_{}\n\n".format(
@@ -415,19 +421,22 @@ def write_probin(probin_template, default_prob_param_file, prob_param_file, out_
 
             elif keyword == "cxx_sets":
                 # this writes out the Fortran functions that can be called from C++
-                # to set the value of the parameters.  Note: we only do this for namelist
-                # parameters
+                # to set the value of the parameters.
 
                 for p in params:
-                    if not p.in_namelist:
+                    if p.is_array() and p.has_module():
                         continue
                     if p.dtype == "logical" or p.dtype == "character":
                         continue
 
                     fout.write("{}subroutine set_f90_{}({}_in) bind(C, name=\"set_f90_{}\")\n".format(
                         indent, p.var, p.var, p.var))
-                    fout.write("{}   {}, intent(in) :: {}_in\n".format(
-                        indent, p.get_f90_decl(), p.var))
+                    if p.is_array():
+                        fout.write("{}   {}, intent(in) :: {}_in({})\n".format(
+                            indent, p.get_f90_decl(), p.var, p.size))
+                    else:
[…]

[…]
