### 2020-09-23 88ea051
Fix case where no parameters use namelist
