### 2020-09-23 5a3cf76
Call probdata_init() for all problem setups
