+++ Exec/hydro_tests/acoustic_pulse/Prob_nd.F90
@@ -12,8 +12,6 @@ subroutine amrex_probinit (init,name,namlen,problo,probhi) bind(c)
   integer,  intent(in) :: name(namlen)
   real(rt), intent(in) :: problo(3), probhi(3)
 
-  call probdata_init(name, namlen)
-
   ! set explosion center
   center(:) = HALF * (problo(:) + probhi(:))
 
