### 2020-09-23 b820e31
Call probdata_init() for all problem setups (#1226)
Rather than rely on problems to manually call probdata_init() in amrex_probinit(), this is now called by default for all setups.

While this has no specific benefit at present (other than simplifying code), it becomes functionally relevant when we start adding parameters to _default_prob_params (#1222).
