+++ Exec/hydro_tests/gresho_vortex/Prob_nd.F90
@@ -18,8 +18,6 @@ subroutine amrex_probinit(init, name, namlen, problo, probhi) bind(c)
   integer,  intent(in   ) :: name(namlen)
   real(rt), intent(in   ) :: problo(3), probhi(3)
 
-  call probdata_init(name, namlen)
-
   ! problem center
   center(:) = (problo(:) + probhi(:)) / 2.e0_rt
 
