+++ Exec/gravity_tests/code_comp/Prob_nd.F90
@@ -15,9 +15,6 @@ subroutine amrex_probinit (init,name,namlen,problo,probhi) bind(c)
   integer, intent(in) :: name(namlen)
   real(rt), intent(in) :: problo(3), probhi(3)
 
-
-  call probdata_init(name, namlen)
-
 #if AMREX_SPACEDIM == 1
   center(1) = ZERO
 
