+++ Exec/science/planet/Prob_nd.F90
@@ -22,8 +22,6 @@ subroutine amrex_probinit(init, name, namlen, problo, probhi) bind(c)
 
   integer :: i
 
-  call probdata_init(name, namlen)
-
   ! Read initial model
   call read_model_file(model_name)
 
