+++ Exec/science/Detonation/Prob_nd.F90
@@ -17,8 +17,6 @@ subroutine amrex_probinit(init, name, namlen, problo, probhi) bind(c)
 
   type(eos_t) :: eos_state
 
-  call probdata_init(name, namlen)
-
   ! get the species indices
   ihe4 = network_species_index("helium-4")
   ic12 = network_species_index("carbon-12")
