+++ Exec/hydro_tests/toy_convect/Prob_nd.F90
@@ -15,9 +15,6 @@ subroutine amrex_probinit(init, name, namlen, problo, probhi) bind(c)
   real(rt) :: offset
   integer :: i
 
-  ! Namelist defaults
-  call probdata_init(name, namlen)
-
   if (num_vortices > max_num_vortices) then
      call castro_error("num_vortices too large, please increase max_num_vortices and the size of xloc_vortices")
   end if
