+++ Exec/gravity_tests/DustCollapse/Prob_nd.F90
@@ -20,8 +20,6 @@ subroutine amrex_probinit(init, name, namlen, problo, probhi) bind(c)
 
   type (eos_t) :: eos_state
 
-  call probdata_init(name, namlen)
-
   r_old_s = r_old
 
   ! in 3-d we center the sphere at (center_x, center_y, center_z)
