### 2020-09-23 5d7219d
Remove call to probdata in Sedov
