### 2020-09-24 9cb3701
Convert center to auto prob params (#1222)
