+++ Exec/science/wdmerger/Castro_ext_src.H
@@ -18,7 +18,6 @@ void do_ext_src(int i, int j, int k,
                 GeometryData const& geomdata, 
                 Array4<const Real> const& state,
                 Array4<Real> const& src,
-                GpuArray<Real, 3> const& center,
                 GpuArray<Real, 3> const& omega,
                 const Real dt, const Real time)
 {
@@ -68,7 +67,7 @@ void do_ext_src(int i, int j, int k,
             for (int dir = 0; dir < AMREX_SPACEDIM; ++dir) {
                 vel[dir] = rhoInv * mom[dir];
             }
-            inertial_to_rotational_velocity(i, j, k, geomdata, center, omega, time, vel);
+            inertial_to_rotational_velocity(i, j, k, geomdata, omega, time, vel);
             for (int dir = 0; dir < AMREX_SPACEDIM; ++dir) {
                 mom[dir] = state(i,j,k,URHO) * vel[dir];
             }
