+++ Source/driver/timestep.cpp
@@ -21,9 +21,6 @@ Castro::estdt_cfl(const Real time)
 
   // Courant-condition limited timestep
 
-  GpuArray<Real, 3> center;
-  ca_get_center(center.begin());
-
 #ifdef ROTATION
   GpuArray<Real, 3> omega;
   get_omega(omega.begin());
@@ -88,7 +85,7 @@ Castro::estdt_cfl(const Real time)
         GeometryData geomdata = geom.data();
 
         inertial_to_rotational_velocity_c(i, j, k, geomdata,
-                                          center.begin(), omega.begin(), time, vel);
+                                          omega.begin(), time, vel);
 
         ux = vel[0];
         uy = vel[1];
