### 2020-09-24 5946103
Add hook for C++ problem initialization
