+++ Source/problems/problem_setup.H
@@ -1,0 +1,5 @@
+AMREX_INLINE
+void problem_initialize () {}
+
+AMREX_GPU_HOST_DEVICE AMREX_INLINE
+void problem_initialize_state_data (int i, int j, int k, Array4<Real> const& state, const GeometryData& geomdata) {}
