### 2020-09-25 cbb25da
Add hook for C++ problem initialization (#1227)
Sedov is converted to the new scheme as a demonstration.
