+++ Exec/hydro_tests/Sedov/Prob_nd.F90
@@ -1,275 +1,0 @@
-subroutine amrex_probinit (init,name,namlen,problo,probhi) bind(c)
-
-  use amrex_fort_module, only: rt => amrex_real
-  use amrex_constants_module, only: ZERO, HALF, ONE, FOUR3RD, M_PI
-  use probdata_module, only: p_ambient, dens_ambient, exp_energy, temp_ambient, e_ambient, &
-                             xn_zone, r_init, nsub, e_exp
-  use prob_params_module, only: center, coord_type
-  use castro_error_module, only: castro_error
-  use eos_type_module, only: eos_t, eos_input_rt, eos_input_rp
-  use eos_module, only: eos
-  use network, only: nspec
-
-  implicit none
-
-  integer,  intent(in) :: init, namlen
-  integer,  intent(in) :: name(namlen)
-  real(rt), intent(in) :: problo(3), probhi(3)
-
-  type(eos_t) :: eos_state
-
-  real(rt) :: vctr
-
-  ! set local variable defaults
-  if (coord_type == 1 .or. coord_type == 2) then
-     center = ZERO
-  else
-     center = HALF * (problo + probhi)
-  end if
-
-  xn_zone(:) = ZERO
-  xn_zone(1) = ONE
-
-  ! override the pressure with the temperature
-  if (temp_ambient > ZERO) then
-
-     eos_state % rho = dens_ambient
-     eos_state % xn(:) = xn_zone(:)
-     eos_state % T = temp_ambient
-
-     call eos(eos_input_rt, eos_state)
-
-     p_ambient = eos_state % p
-
-  endif
-
-  ! Calculate ambient state data
-
-  eos_state % rho = dens_ambient
-  eos_state % p   = p_ambient
-  eos_state % T   = 1.e9_rt ! Initial guess for iterations
-  eos_state % xn  = xn_zone
-
-  call eos(eos_input_rp, eos_state)
-
-  e_ambient = eos_state % e
-  temp_ambient = eos_state % T
-
-  ! set explosion pressure -- we will convert the point-explosion energy into
-  ! a corresponding pressure distributed throughout the perturbed volume
-
-  if (coord_type == 0) then
-
-#if AMREX_SPACEDIM == 1
-
-#ifndef AMREX_USE_CUDA
-     call castro_error("Sedov problem unsupported in 1D Cartesian geometry.")
-#endif
-
-#elif AMREX_SPACEDIM == 2
-
-     ! Cylindrical problem in Cartesian coordinates
-
-     vctr = M_PI*r_init**2
-
-#else
-
-     ! Spherical problem in Cartesian coordinates
-
-     vctr = FOUR3RD*M_PI*r_init**3
-
-#endif
-
-  else if (coord_type == 1) then
-
-#if AMREX_SPACEDIM == 1
-
-     vctr = M_PI*r_init**2
-
-#elif AMREX_SPACEDIM == 2
-
-     vctr = FOUR3RD*M_PI*r_init**3
-
-#else
-
-#ifndef AMREX_USE_CUDA
-     call castro_error("Sedov problem unsupported in 3D axisymmetric geometry.")
-#endif
-
-#endif
-
-  else if (coord_type == 2) then
-
-     ! Must have AMREX_SPACEDIM == 1 for this coord_type.
-
-     vctr = FOUR3RD*M_PI*r_init**3
-
-  end if
-
-  e_exp = exp_energy/vctr/dens_ambient
-
-end subroutine amrex_probinit
-
-
-
-subroutine ca_initdata(lo, hi, &
-                       state, s_lo, s_hi, &
-                       dx, problo) bind(C, name='ca_initdata')
-
[…]
