### 2020-09-25 2fa1807
convert this problem setup to C++
