+++ Exec/hydro_tests/acoustic_pulse/problem_setup.H
@@ -56,7 +56,9 @@ void problem_initialize_state_data (int i, int j, int k,
         dist = std::abs(center[2] - zz);
 
     } else {
+#ifndef AMREX_USE_GPU 
         amrex::Error("invalid init_as_1d");
+#endif
     }
 
     if (dist <= 0.5_rt) {
