### 2020-09-25 a62d9b6
convert the acoustic_pulse problem setup to C++ (#1228)
