### 2020-09-25 cc05734
update to use the NSE tables this is now based off of reacting_convergence
