### 2020-09-26 f0deebe
Add prob_params variables to a C++ namespace
