+++ Source/gravity/Castro_pointmass.cpp
@@ -40,14 +40,14 @@ Castro::pointmass_update(Real time, Real dt)
                 const Real eps = 1.e-8_rt;
 
                 // This should be the cell whose lower left corner is at center
-                int icen = std::floor((center[0] - problo[0]) / dx[0] + eps);
+                int icen = std::floor((problem::center[0] - problo[0]) / dx[0] + eps);
 #if AMREX_SPACEDIM >= 2
-                int jcen = std::floor((center[1] - problo[1]) / dx[1] + eps);
+                int jcen = std::floor((problem::center[1] - problo[1]) / dx[1] + eps);
 #else
                 int jcen = 0;
 #endif
 #if AMREX_SPACEDIM == 3
-                int kcen = std::floor((center[2] - problo[2]) / dx[2] + eps);
+                int kcen = std::floor((problem::center[2] - problo[2]) / dx[2] + eps);
 #else
                 int kcen = 0;
 #endif
@@ -116,14 +116,14 @@ Castro::pointmass_update(Real time, Real dt)
                     const Real eps = 1.e-8_rt;
 
                     // This should be the cell whose lower left corner is at center
-                    int icen = std::floor((center[0] - problo[0]) / dx[0] + eps);
+                    int icen = std::floor((problem::center[0] - problo[0]) / dx[0] + eps);
 #if AMREX_SPACEDIM >= 2
-                    int jcen = std::floor((center[1] - problo[1]) / dx[1] + eps);
+                    int jcen = std::floor((problem::center[1] - problo[1]) / dx[1] + eps);
 #else
                     int jcen = 0;
 #endif
 #if AMREX_SPACEDIM == 3
-                    int kcen = std::floor((center[2] - problo[2]) / dx[2] + eps);
+                    int kcen = std::floor((problem::center[2] - problo[2]) / dx[2] + eps);
 #else
                     int kcen = 0;
 #endif
