+++ Source/hydro/Castro_ctu_hydro.cpp
@@ -1257,7 +1257,7 @@ Castro::construct_ctu_hydro_source(Real time, Real dt)
           position(i, j, k, geomdata, loc);
 
           for (int dir = 0; dir < AMREX_SPACEDIM; ++dir)
-              loc[dir] -= center[dir];
+              loc[dir] -= problem::center[dir];
 
           Real R = amrex::max(std::sqrt(loc[0] * loc[0] + loc[1] * loc[1]), R_min);
           Real RInv = 1.0_rt / R;
