### 2020-09-26 4f94feb
Fix various problem setups
