### 2020-09-26 5bc1d8c
add a convergence testing script
