### 2020-09-26 0d1c846
Add prob_params variables to a C++ namespace (#1232)
Variables are added to the problem namespace.
