### 2020-09-27 fa99c12
convert this to C++
