### 2020-09-27 1a775ce
convert reacting_convergence to C++ (#1234)
