+++ Exec/hydro_tests/Sod/README.md
@@ -1,0 +1,8 @@
+# Sod
+
+This is a one-dimensional shock tube problem that is used to run the
+classic Sod problem.  inputs/probin files are also provided to run
+most of the test problems defined in Toro's text.
+
+Plotting against the analytic solution can be done using the scripts
+and data in `Verification/`.
