### 2020-09-27 d2fdee0
switch the code_comp derives over to C++
