### 2020-09-27 85a3b52
switch the gamma_law_bubble derives over to C++
