### 2020-09-27 d73e52f
convert diffusion_test derives to C++
