### 2020-09-27 b4bab54
convert RadSphere derives to C++
