### 2020-09-27 1d9d10b
convert the RadThermalWave derives to C++
