### 2020-09-27 539e18b
convert Sod to C++ (#1235)
