### 2020-09-27 a7140ad
fix loc accesses
