### 2020-09-27 3722a8b
convert the RadThermalWave derives to C++ (#1240)
