### 2020-09-27 f3eb172
guard against invalid accesses
