+++ Exec/unit_tests/diffusion_test/Problem_Derive.cpp
@@ -27,10 +27,14 @@ void deranalytic(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
   [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
   {
 
-    Real r[3];
+    Real r[3] = {0.0};
     r[0] = problo[0] + dx[0] * (static_cast<Real>(i) + 0.5_rt);
+#if AMREX_SPACEDIM >= 2
     r[1] = problo[1] + dx[1] * (static_cast<Real>(j) + 0.5_rt);
+#endif
+#if AMREX_SPACEDIM == 3
     r[2] = problo[2] + dx[2] * (static_cast<Real>(k) + 0.5_rt);
+#endif
 
     Real exponent = 0.0;
 
