### 2020-09-27 0c3b6de
switch the gamma_law_bubble derives over to C++ (#1237)
