+++ Exec/hydro_tests/gamma_law_bubble/Problem_Derive.cpp
@@ -1,99 +1,232 @@
 #include <AMReX_REAL.H>
 
 #include <Derive.H>
-#include <Problem_Derive_F.H>
 #include <Castro.H>
 #include <Castro_F.H>
+#include <prob_util.H>
 
 using namespace amrex;
 
-#ifdef __cplusplus
-extern "C"
+using RealVector = amrex::Gpu::ManagedVector<amrex::Real>;
+
+void ca_derpi(const Box& bx, FArrayBox& derfab, int dcomp, int /*ncomp*/,
+              const FArrayBox& datfab, const Geometry& geomdata,
+              Real /*time*/, const int* /*bcrec*/, int /*level*/)
 {
+
+  // derive the dynamic pressure
+
+  const auto dx = geomdata.CellSizeArray();
+  const auto problo = geomdata.ProbLoArray();
+
+  auto const dat = datfab.array();
+  auto const der = derfab.array();
+
+  // first make a 1D initial model for the entire domain
+  int npts_1d = (2.0_rt * problem::center[AMREX_SPACEDIM-1] + 1.e-8_rt) /
+    dx[AMREX_SPACEDIM-1];
+
+  RealVector pressure_rv(npts_1d, 0);
+  RealVector density_rv(npts_1d, 0);
+  RealVector temp_rv(npts_1d, 0);
+  RealVector eint_rv(npts_1d, 0);
+
+  Real* const pressure = pressure_rv.dataPtr();
+  Real* const density = density_rv.dataPtr();
+  Real* const temp = temp_rv.dataPtr();
+  Real* const eint = eint_rv.dataPtr();
+
+  gamma_law_initial_model(pressure, density, temp, eint, npts_1d, dx);
+
+  // Compute pressure from the EOS
+
+  amrex::ParallelFor(bx,
+  [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k) noexcept
+  {
+
+    Real rhoInv = 1.0_rt / dat(i,j,k,URHO);
+    Real T = dat(i,j,k,UTEMP);
+
+    eos_t eos_state;
+
+    eos_state.rho = dat(i,j,k,URHO);
+    eos_state.T = dat(i,j,k,UTEMP);
+    eos_state.e = dat(i,j,k,UEINT) * rhoInv;
+    for (int n = 0; n < NumSpec; n++) {
+      eos_state.xn[n] = dat(i,j,k,UFS+n) / dat(i,j,k,URHO);
+    }
+#if NAUX > 0
+    for (int n = 0; n < NumAux; n++) {
+      eos_state.aux[n] = dat(i,j,k,UFX+n) / dat(i,j,k,URHO);
+    }
 #endif
 
-    // Note that in the following routines, we are NOT passing
-    // several variables to Fortran that would be unused.
+    if (eos_state.e <= 0.0_rt) {
+      eos(eos_input_rt, eos_state);
+      der(i,j,k,0) = eos_state.p;
 
-    // These routines are called in an MFIter loop, so we do not
-    // need to explicitly synchronize after GPU kernels.
-
-    void ca_derpi(Real* der, const int* der_lo, const int* der_hi, const int* nvar,
-                  const Real* data, const int* data_lo, const int* data_hi, const int* ncomp,
-                  const int* lo, const int* hi,
-                  const int* domain_lo, const int* domain_hi,
-                  const Real* dx, const Real* xlo,
-                  const Real* time, const Real* dt, const int* bcrec, 
-                  const int* level, const int* grid_no)
-    {
-
-#pragma gpu
-        derpi(der, AMREX_INT_ANYD(der_lo), AMREX_INT_ANYD(der_hi), *nvar,
[…]
