### 2020-09-28 8b6b7d4
set the composition if we are in NSE appropriately
