### 2020-09-28 0f2590a
set the composition if we are in NSE appropriately (#1241)
