### 2020-09-29 2e5afbd
Implement new AMReX tagging format (#1243)
The refinement scheme from AMReX-Codes/amrex#1166 is implemented. Rather than using the Fortran tagging namelist, applications may use the inputs file. The scheme looks like:
```
amr.refinement_indicators = dens temp

amr.refine.dens.max_level = 1
amr.refine.dens.value_greater = 2.0
amr.refine.dens.field_name = Density

amr.refine.temp.max_level = 2
amr.refine.temp.value_less = 1.0
amr.refine.temp.field_name = Temp
```
amr.refinement_indicators is a list of user-defined names for refinement schemes. For each defined name, amr.refine.<name> accepts pre
[…]
