+++ Source/driver/Castro.H
@@ -1181,6 +1181,11 @@ public:
     static amrex::Vector<int> qpass_map;
 
 ///
+/// Vector storing list of custom error tags.
+///
+    static amrex::Vector<amrex::AMRErrorTag> custom_error_tags;
+
+///
 /// This MultiFab is on the coarser level.  This is useful for the coarser level
 ///     to mask out the finer level.  We only build this when it is needed.
 ///     This coarse MultiFab has to live on the fine level because it must be updated
