+++ CHANGES.md
@@ -1,14 +1,14 @@
 # 20.10
 
    * A new refinement scheme using the inputs file rather than the Fortran
-     tagging namelist has been added. (#1243) As an example, consider:
+     tagging namelist has been added. (#1243, #1245) As an example, consider:
 
      ```
      amr.refinement_indicators = dens temp
 
      amr.refine.dens.max_level = 1
      amr.refine.dens.value_greater = 2.0
-     amr.refine.dens.field_name = Density
+     amr.refine.dens.field_name = density
 
      amr.refine.temp.max_level = 2
      amr.refine.temp.value_less = 1.0
@@ -20,10 +20,10 @@
      describing when to tag. In the current implementation, these are `max_level`
      (maximum level to refine to), `start_time` (when to start tagging), `end_time`
      (when to stop tagging), `value_greater` (value above which we refine),
-     `value_less` (value below which to refine), and `field_name` (name of the
-     string defining the field in the code). If a refinement indicator is added,
-     either `value_greater` or `value_less` must be provided. (In the future we
-     will support other refinement schemes such as gradients.)
+     `value_less` (value below which to refine), `gradient` (absolute value of the
+     difference between adjacent cells above which we refine), and `field_name`
+     (name of the string defining the field in the code). If a refinement indicator
+     is added, either `value_greater`, `value_less`, or `gradient` must be provided.
 
    * Automatic problem parameter configuration is now available to every
      problem by placing a _prob_params file in your problem directory.
