### 2020-09-29 45db979
Add gradient to inputs-based error tagging (#1246)
Sedov is converted to the new scheme as a demonstration.
