+++ Util/model_parser_cxx/model_parser.H
@@ -1,0 +1,297 @@
+#ifndef MODEL_PARSER_H
+#define MODEL_PARSER_H
+
+#include <iostream>
+#include <fstream>
+#include <vector>
+#include <algorithm>
+
+
+///
+/// read in an initial model and return arrays with the model data.
+/// take care to match the species available in the model file to
+/// those defined by the network
+///
+/// the model file is assumed to be of the follow form:
+/// # npts = 896
+/// # num of variables = 6
+/// # density
+/// # temperature
+/// # pressure
+/// # carbon-12
+/// # oxygen-16
+/// # magnesium-24
+/// 195312.5000  5437711139.  8805500.952   .4695704813E+28  0.3  0.7  0
+/// 585937.5000  5410152416.  8816689.836  0.4663923963E+28  0.3  0.7  0
+///
+/// we read in the number of variables and their order and use this to
+/// map them into the model_state array.  We ignore anything other than
+/// density, temperature, pressure and composition.
+///
+/// composition is assumed to be in terms of mass fractions
+
+
+namespace model
+{
+
+    // integer keys for indexing the model_state array
+
+    constexpr int nvars = 3 + NumSpec;
+    constexpr int idens = 0;
+    constexpr int itemp = 1;
+    constexpr int ipres = 2;
+    constexpr int ispec = 3;
+
+    AMREX_GPU_MANAGED int npts;
+    AMREX_GPU_MANAGED bool initialized;
+
+    amrex::Gpu::ManagedVector<Real> state;
+    amrex::Gpu::ManagedVector<Real> r;
+}
+
+#define MODELINDEX(I, J) model::npts*I+J;
+
+
+AMREX_INLINE
+void
+read_model_file(std::string& model_file) {
+
+    bool found_model, found_dens, found_temp, found_pres;
+    bool found_spec[NumSpec];
+
+    // read in the initial model
+
+    std::ifstream initial_model_file;
+
+    initial_model_file.open("model.hse.cool.coulomb", std::ios::in);
+    if (!initial_model_file.is_open()) {
+        amrex::Error("Error opening the initial model");
+    }
+
+    std::string line;
+
+    // first the header line -- this tells us the number of points
+
+    getline(initial_model_file, line);
+    std::string npts_string = line.substr(line.find("=")+1, line.length());
+    model::npts = std::stoi(npts_string);
+
+    // next line tells use the number of variables
+
+    getline(initial_model_file, line);
+    std::string num_vars_string = line.substr(line.find("=")+1, line.length());
+    int nvars_model_file = std::stoi(num_vars_string);
+
+    // now read in the names of the variables
+
+    std::vector<std::string> varnames_stored;
+    for (int n = 0; n < nvars_model_file; n++) {
+        getline(initial_model_file, line);
+        std::string var_string = line.substr(line.find("#")+1, line.length());
+        varnames_stored.push_back(ltrim(var_string));
+    }
+
+
+    // allocate storage for the model data
+
+    model::state.resize(model::npts * model::nvars);
+    model::r.resize(npts_model);
+
+    amrex::Print() << "reading initial model" << std::endl;
+    amrex::Print() << npts_model << " points found in the initial model" << std::endl;
+    amrex::Print() << nvars_model_file << " variables found in the initial model file" << std::endl;
+
+    // start reading in the data
+    Real vars_stored[nvars_model_file];
+
+    for (int i = 0; i < model::npts; i++) {
+        initial_model_file >> model::r[i];
+
+        for (int j = 0; j < nvars_model_file; j++) {
+            initial_model_file >> vars_stored[j];
+        }
+
+        for (int j = 0; j < nvars_model; j++) {
+            model::state(MODELINDEX(i,j)) = 0.0_rt;
[…]
