### 2020-09-30 a64e4a4
start converting this over to C++
