+++ Exec/hydro_tests/toy_convect/Prob_nd.F90
@@ -1,184 +1,0 @@
-subroutine amrex_probinit(init, name, namlen, problo, probhi) bind(c)
-
-  use probdata_module
-  use model_parser_module
-  use castro_error_module
-  use amrex_fort_module, only : rt => amrex_real
-  use amrex_paralleldescriptor_module, only: parallel_IOProcessor => amrex_pd_ioprocessor
-
-  implicit none
-
-  integer,  intent(in) :: init, namlen
-  integer,  intent(in) :: name(namlen)
-  real(rt), intent(in) :: problo(3), probhi(3)
-
-  real(rt) :: offset
-  integer :: i
-
-  if (num_vortices > max_num_vortices) then
-     call castro_error("num_vortices too large, please increase max_num_vortices and the size of xloc_vortices")
-  end if
-
-  ! Read initial model
-  call read_model_file(model_name)
-
-
-  if (parallel_IOProcessor()) then
-     do i = 1, npts_model
-        print *, i, model_r(i), model_state(i,idens_model)
-     enddo
-  endif
-
-  ! velocity perturbation stuff
-  offset = (probhi(1) - problo(1)) / (num_vortices)
-
-  do i = 1, num_vortices
-     xloc_vortices(i) = (dble(i-1) + 0.5e0_rt) * offset + problo(1)
-  enddo
-
-  ! store the keys for the source term later
-  ih1 = network_species_index("hydrogen-1")
-  ic12 = network_species_index("carbon-12")
-  in14 = network_species_index("nitrogen-14")
-  io16 = network_specieS_index("oxygen-16")
-
-end subroutine amrex_probinit
-
-
-! ::: -----------------------------------------------------------
-! ::: This routine is called at problem setup time and is used
-! ::: to initialize data on each grid.
-! :::
-! ::: NOTE:  all arrays have one cell of ghost zones surrounding
-! :::        the grid interior.  Values in these cells need not
-! :::        be set here.
-! :::
-! ::: INPUTS/OUTPUTS:
-! :::
-! ::: level     => amr level of grid
-! ::: time      => time at which to init data
-! ::: lo,hi     => index limits of grid interior (cell centered)
-! ::: nstate    => number of state components.  You should know
-! :::		   this already!
-! ::: state     <=  Scalar array
-! ::: delta     => cell size
-! ::: xlo,xhi   => physical locations of lower left and upper
-! :::              right hand corner of grid.  (does not include
-! :::		   ghost region).
-! ::: -----------------------------------------------------------
-subroutine ca_initdata(level, time, lo, hi, nscal, &
-                       state, s_lo, s_hi, &
-                       delta, xlo, xhi)
-
-  use probdata_module
-  use eos_module
-  use eos_type_module, only : eos_t, eos_input_rt
-  use meth_params_module, only : NVAR, URHO, UMX, UMY, UMZ, UEDEN, UEINT, UFS, UTEMP
-  use network, only: nspec
-  use model_parser_module
-  use amrex_constants_module, only : ZERO, HALF
-  use amrex_fort_module, only : rt => amrex_real
-  use prob_params_module, only : problo
-
-  implicit none
-
-  integer,  intent(in   ) :: level, nscal
-  integer,  intent(in   ) :: lo(3), hi(3)
-  integer,  intent(in   ) :: s_lo(3), s_hi(3)
-  real(rt), intent(in   ) :: xlo(3), xhi(3), time, delta(3)
[…]
