+++ Util/model_parser_cxx/model_parser.H
@@ -30,6 +30,13 @@
 ///
 /// composition is assumed to be in terms of mass fractions
 
+// remove whitespace -- from stackoverflow
+static inline std::string &ltrim(std::string &s) {
+    s.erase(s.begin(), std::find_if(s.begin(), s.end(), [](int c) {return !std::isspace(c);}));
+    return s;
+}
+
+
 
 namespace model
 {
@@ -49,8 +56,111 @@ namespace model
     amrex::Gpu::ManagedVector<Real> r;
 }
 
-#define MODELINDEX(I, J) model::npts*I+J;
+#define MODELINDEX(I, J) model::npts*I+J
 
+
+
+AMREX_INLINE AMREX_GPU_HOST_DEVICE
+int
+locate(const Real r) {
+
+    int loc;
+
+    if (r <= model::r[0]) {
+       loc = 0;
+
+    } else if (r > model::r[model::npts-2]) {
+       loc = model::npts-1;
+
+    } else {
+
+        int ilo = 0;
+        int ihi = model::npts-2;
+
+        while (ilo+1 != ihi) {
+            int imid = (ilo + ihi) / 2;
+
+            if (r <= model::r[imid]) {
+                ihi = imid;
+            } else {
+                ilo = imid;
+            }
+        }
+
+        loc = ihi;
+    }
+
+    return loc;
+}
+
+
+AMREX_INLINE AMREX_GPU_HOST_DEVICE
+Real
+interpolate(const Real r, const int var_index) {
+
+    // find the value of model_state component var_index at point r
+    // using linear interpolation.  Eventually, we can do something
+    // fancier here.
+
+    int id = locate(r);
+
+    Real slope;
+    Real interp;
+
+    if (id == 0) {
+
+       slope = (model::state[MODELINDEX(id+1, var_index)] -
+                model::state[MODELINDEX(id, var_index)]) /
+           (model::r[id+1] - model::r[id]);
+       interp = slope * (r - model::r[id]) + model::state[MODELINDEX(id, var_index)];
+
+       // safety check to make sure interp lies within the bounding points
+       Real minvar = amrex::min(model::state[MODELINDEX(id+1, var_index)],
+                                model::state[MODELINDEX(id, var_index)]);
+       Real maxvar = amrex::max(model::state[MODELINDEX(id+1, var_index)],
+                                model::state[MODELINDEX(id, var_index)]);
+       interp = amrex::max(interp, minvar);
+       interp = amrex::min(interp, maxvar);
+
+    } else if (id == model::npts-1) {
+
+       slope = (model::state[MODELINDEX(id, var_index)] -
+                model::state[MODELINDEX(id-1, var_index)]) /
+           (model::r[id] - model::r[id-1]);
+       interp = slope * (r - model::r[id]) + model::state[MODELINDEX(id, var_index)];
+
+
+       // safety check to make sure interp lies within the bounding points
+       Real minvar = amrex::min(model::state[MODELINDEX(id-1, var_index)],
+                                model::state[MODELINDEX(id, var_index)]);
+       Real maxvar = amrex::max(model::state[MODELINDEX(id-1, var_index)],
+                                model::state[MODELINDEX(id, var_index)]);
+       interp = amrex::max(interp, minvar);
+       interp = amrex::min(interp, maxvar);
+
+    } else {
+
+       if (r >= model::r[id]) {
+
+          slope = (model::state[MODELINDEX(id+1, var_index)] -
+                   model::state[MODELINDEX(id, var_index)]) /
+              (model::r[id+1] - model::r[id]);
+          interp = slope * (r - model::r[id]) + model::state[MODELINDEX(id, var_index)];
+
+       } else {
+
[…]
