+++ Util/model_parser_cxx/model_parser_data.cpp
@@ -1,5 +1,5 @@
 #include <AMReX_Vector.H>
-#include "nse.H"
+#include <AMReX_GpuContainers.H>
 
 namespace model
 {
@@ -7,7 +7,7 @@ namespace model
     extern AMREX_GPU_MANAGED int npts;
     extern AMREX_GPU_MANAGED bool initialized;
 
-    extern amrex::Gpu::ManagedVector<Real> state;
-    extern amrex::Gpu::ManagedVector<Real> r;
+    extern amrex::Gpu::ManagedVector<amrex::Real> state;
+    extern amrex::Gpu::ManagedVector<amrex::Real> r;
 
 }
