+++ Source/problems/Castro_bc_ext_fill_nd.H
@@ -9,9 +9,11 @@ void ca_ext_fill(amrex::Box const& bx, amrex::FArrayBox& data,
                  amrex::Geometry const& geom, const amrex::Real time,
                  const int* bc_f);
 
+#ifndef CXX_MODEL_PARSER
 void ca_ext_denfill(amrex::Box const& bx, amrex::FArrayBox& data,
                     const int dcomp, const int numcomp,
                     amrex::Geometry const& geom, const amrex::Real time,
                     const int* bc_f);
+#endif
 
 #endif
