+++ Util/model_parser_cxx/model_parser_data.cpp
@@ -1,5 +1,6 @@
 #include <AMReX_Vector.H>
 #include <AMReX_GpuContainers.H>
+#include <model_parser.H>
 
 namespace model
 {
