### 2020-09-30 dc3df99
fix some build issues
