### 2020-10-01 3b88aa3
we were not updating the kinetic energy to account for the perturbation
