### 2020-10-01 747b326
changes made to mhd burning (perturbation added)
