### 2020-10-02 ba93682
convert RT to C++
