+++ Exec/gravity_tests/hse_convergence/initial_model.H
@@ -1,0 +1,341 @@
+// create a simple hydrostatic initial model that is 4th order accurate
+// at cell-centers by using RK4 to integrate:
+//
+//   dp/dr = rho(p, s) g
+//   ds/dr = 0
+
+
+struct model_t {
+    Real dens_base = -1;
+    Real T_base = -1;
+    Real xn[NumSpec] = {0.0};
+}
+
+
+///
+/// construct an initial model in HSE.  Note: this does not return
+/// anything, but rather updates the model_parser globals with the
+/// model information.
+///
+AMREX_INLINE
+void
+generate_initial_model(const int nx, const Real xmin, const Real xmax,
+                       const model_t model_params, const int nbuf) {
+
+    constexpr int MAX_ITER = 100;
+    constexpr Real TOL = 1.e-10_rt;
+
+    // allocate the storage in the model_parser_module
+    int npts_model = nx + 2*nbuf;
+
+    allocate_model_parser_data(npts_model);
+
+    int ibase = nbuf;
+    int itop = ibase + nx - 1;
+
+    // get the base conditions
+
+    eos_t eos_state;
+    eos_state.rho = model_params.dens_base;
+    eos_state.T = model_params.T_base;
+    for (int n = 0; n < NumSpec; n++) {
+        eos_state.xn[n] = model_params.xn[n];
+    }
+
+    eos(eos_input_rt, eos_state);
+
+    Real entropy_fixed = eos_state.s;
+
+    // create the grid -- cell centers
+
+    Real dx = (xmax - xmin) / nx;
+
+    for (int i = 0; i < npts_model; i++) {
+        model::r[i] = xmin + (static_cast<Real>(i - nbuf) + 0.5_rt) * dx;
+    }
+
+    // note, those conditions are the lower boundary.  This means we
+    // need to integrate dx/2 in the first step to get to the first
+    // zone's cell-center.  After that, we integrate from
+    // center-to-center.
+
+    Real p = eos_state.p;
+
+    if (time_integration_method == 2 && sdc_order == 4) {
+
+        // a fourth order accurate method
+
+        // do RK 4 integration up from the lower boundary
+
+        for (int i = ibase; i <= itop + nbuf; i++) {
+
+            // rho and T here are guesses for the EOS call
+
+            Real rho;
+            Real T;
+
+            if (i == ibase) {
+                rho = eos_state.rho;
+                T = eos_state.T;
+            } else {
+                rho = model::state(MODELINDEX(i-1, model::idens));
+                T = model::state(MODELINDEX(i-1, model::itemp));
+            }
+
+            Real xn[NumSpec];
+            for (int n = 0; n < NumSpec; n++) {
+                xn[n] = model_params.xn[n];
+            }
+
+            // step size
+
+            Real h;
+            if (i == ibase) {
+                h = 0.5_rt * dx;
+            } else {
+                h = dx;
+            }
+
+            // entropy never changes in this model
+
+            Real s = entropy_fixed;
+
+            Real k1 = f(p, s, const_grav, rho, T, xn);
+            Real k2 = f(p + 0.5_rt * h * k1, s, const_grav, rho, T, xn);
+            Real k3 = f(p + 0.5_rt * h * k2, s, const_grav, rho, T, xn);
+            Real k4 = f(p + h * k3, s, const_grav, rho, T, xn);
+
+            Real pnew = p + (1.0_rt/6.0_rt) * h * (k1 + 2.0_rt * k2 + 2.0_rt * k3 + k4);
+
+            // call the EOS to get the remainder of the thermodynamics
+
+            eos_t eos_state;
+            eos_state.T = T; // initial guess
+            eos_state.rho = rho; // initial guess
[…]
