+++ Exec/gravity_tests/hse_convergence/problem_setup.H
@@ -1,0 +1,110 @@
+#ifndef problem_setup_H
+#define problem_setup_H
+
+#include <prob_parameters.H>
+#include <eos.H>
+#include <model_parser.H>
+#include <initial_model.H>
+#include <network.H>
+
+AMREX_INLINE
+void problem_initialize ()
+{
+    const Geometry& dgeom = DefaultGeometry();
+
+    const Real* problo = dgeom.ProbLo();
+    const Real* probhi = dgeom.ProbHi();
+
+    for (int n = 0; n < AMREX_SPACEDIM; ++n) {
+        problem::center[n] = 0.5_rt * (problo[n] + probhi[n]);
+    }
+
+
+    int ihe4 = network_spec_index("helium-4");
+    if (ihe4 < 0) {
+        amrex::Error("Error: helium-4 not present");
+    }
+
+    model_t model_params;
+    model_params.T_base = problem::temp_base;
+    model_params.dens_base = problem::dens_base;
+    for (int n = 0; n < NumSpec; n++) {
+        model_params.xn[n] = 100*small_x;
+    }
+    model_params.xn[ihe4] = 1.0_rt - (NumSpec - 1) * 100*small_x;
+
+    // we add some buffer to the model so we can use it to fill ghost
+    // cells in the boundary conditions
+
+    int nbuf = 8;
+
+    // generate the initial model -- it will be stored in the model
+    // parser global data
+
+    generate_initial_model(problem::nx_model,
+                           problo[AMREX_SPACEDIM-1], probhi[AMREX_SPACEDIM-1],
+                           model_params, nbuf);
+
+}
+
+
+AMREX_GPU_HOST_DEVICE AMREX_INLINE
+void problem_initialize_state_data (int i, int j, int k,
+                                    Array4<Real> const& state,
+                                    const GeometryData& geomdata)
+{
+    int coord_type = geomdata.Coord();
+
+    const Real* dx = geomdata.CellSize();
+    const Real* problo = geomdata.ProbLo();
+
+    Real x = problo[0] + dx[0] * (static_cast<Real>(i) + 0.5_rt);
+
+    Real y = 0.0;
+#if AMREX_SPACEDIM >= 2
+    y = problo[1] + dx[1] * (static_cast<Real>(j) + 0.5_rt);
+#endif
+
+    Real z = 0.0;
+#if AMREX_SPACEDIM == 3
+    z = problo[2] + dx[2] * (static_cast<Real>(k) + 0.5_rt);
+#endif
+
+#if AMREX_SPACEDIM == 1
+    Real height = x;
+#elif AMREX_SPACEDIM == 2
+    Real height = y;
+#else
+    Real height = z;
+#endif
+
+    state(i,j,k,URHO) = interpolate(height, model::idens);
+    state(i,j,k,UTEMP) = interpolate(height, model::itemp);
+    for (int n = 0; n < NumSpec; n++) {
+        state(i,j,k,UFS+n) = interpolate(height, model::ispec+n);
+    }
+
+    eos_t eos_state;
+    eos_state.rho = state(i,j,k,URHO);
+    eos_state.T = state(i,j,k,UTEMP);
+    for (int n = 0; n < NumSpec; n++) {
+        eos_state.xn[n] = state(i,j,k,UFS+n);
+    }
+
+    eos(eos_input_rt, eos_state);
+
+    state(i,j,k,UEINT) = state(i,j,k,URHO) * eos_state.e;
+    state(i,j,k,UEDEN) = state(i,j,k,UEINT);
+
+    for (int n = 0; n < NumSpec; n++) {
[…]
