### 2020-10-05 753f36e
move problem_initialize() to the end of the constructor this way we can be sure all of the C++ ParmParse's were done before the problem starts
