### 2020-10-05 9fb3045
move all of the runtime parameter initialization to a central place
