### 2020-10-06 673879a
set initial guesses
