### 2020-10-06 e8b2ff8
move the radsolve inits back into that class
