### 2020-10-07 f6d891c
for some reason we were disabling the ext denfill BCs for the x-direction on GPUs (#1248)
