### 2020-10-08 2275df3
move all of the runtime parameter initialization to a central place (#1257)
this moves all of the parmparse stuff for the _cpp_parameter initialization into a single initialization routine (except for the radsolve parameters for now).  This change allows the C++ initialization to use the values of other runtime parameters, like gravity.
