### 2020-10-13 a45e14b
Permit level dependent thresholds in new tagging scheme
