+++ Source/driver/Castro.cpp
@@ -500,23 +500,23 @@ Castro::read_params ()
             info.SetMaxLevel(max_level);
         }
 
-        if (ppr.countval("value_greater") > 0) {
-            Real value;
-            ppr.get("value_greater", value);
+        if (int nval = ppr.countval("value_greater")) {
+            Vector<Real> value;
+            ppr.getarr("value_greater", value, 0, nval);
             std::string field;
             ppr.get("field_name", field);
             custom_error_tags.push_back(AMRErrorTag(value, AMRErrorTag::GREATER, field, info));
         }
-        else if (ppr.countval("value_less") > 0) {
-            Real value;
-            ppr.get("value_less", value);
+        else if (int nval = ppr.countval("value_less")) {
+            Vector<Real> value;
+            ppr.getarr("value_less", value, 0, nval);
             std::string field;
             ppr.get("field_name", field);
             custom_error_tags.push_back(AMRErrorTag(value, AMRErrorTag::LESS, field, info));
         }
-        else if (ppr.countval("gradient") > 0) {
-            Real value;
-            ppr.get("gradient", value);
+        else if (int nval = ppr.countval("gradient")) {
+            Vector<Real> value;
+            ppr.getarr("gradient", value, 0, nval);
             std::string field;
             ppr.get("field_name", field);
             custom_error_tags.push_back(AMRErrorTag(value, AMRErrorTag::GRAD, field, info));
