### 2020-10-13 #2513 @zingale
Here's a Zenodo DOI for a tagged release:
https://zenodo.org/record/4085328
