### 2020-10-13 93ade34
Permit level dependent thresholds in new tagging scheme (#1260)
We take advantage of AMReX-Codes/amrex#1441 to allow the user to specify a tagging threshold that can vary by level. If a value for every level is not provided, we assume the last value holds for all remaining levels, similar to how we treat quantities like amr.n_error_buf.
