### 2020-10-13 #2513 @eloisabentivegna
@zingale, I am wondering whether the release should have a traditional numerical tag, rather than a special label like "JOSS_paper", so users can readily identify whether a given code version precedes or follows the one published in JOSS.

@arfon, @danielskatz, what do you think?
