### 2020-10-14 #2513 @arfon
We don't enforce anything along those lines here and are happy to follow whatever conventions a project might already have in place.
