### 2020-10-14 #2513 @eloisabentivegna
> We don't enforce anything along those lines here and are happy to follow whatever conventions a project might already have in place.

OK! The project follows a mix of the usual convention (numerical) and special descriptive tags. If there's no requirement on our side, let's leave everything as is. Thanks!
