### 2020-10-14 2a12e43
put the string methods in a namespace
