### 2020-10-14 bff8edb
Using inputs-det-x.nse file
