### 2020-10-16 1331028
Fix missing/invalid DOIs in paper.bib.
