+++ paper/paper.bib
@@ -3,7 +3,15 @@
   author={D. Fan and A. Nonaka and A. S. Almgren and A. Harpole and M. Zingale},
   journal={accepted for publication in The Astrophysical Journal},
   year={2019},
-  url = {https://arxiv.org/abs/1908.03634}
+  journal = {\apj},
+  year = 2019,
+  volume = {887},
+  number = {2},
+  eid = {212},
+  pages = {212},
+  archivePrefix = {arXiv},
+  eprint = {1908.03634},
+  doi = {10.3847/1538-4357/ab4f75}
 }
 
 @article{AMReX,
@@ -153,6 +161,11 @@ archivePrefix = {arXiv},
 archivePrefix = {arXiv},
        eprint = {1910.12578},
  primaryClass = {astro-ph.IM},
+    publisher = {{IOP} Publishing},
+       volume = {1623},
+	pages = {012021},
+      journal = {Journal of Physics: Conference Series},
+          doi = {10.1088/1742-6596/1623/1/012021},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv191012578Z},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
@@ -204,6 +217,12 @@ archivePrefix = {arXiv},
 archivePrefix = {arXiv},
        eprint = {1905.04341},
  primaryClass = {cs.DC},
+      journal = {IEEE Transactions on Parallel and Distributed Systems},
+         year = {2021},
+       volume = {32},
+       number = {1},
+        pages = {85-97},
+          doi = {10.1109/tpds.2020.3010016},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190504341G},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
@@ -292,7 +311,7 @@ journal = "Journal of Parallel and Distributed Computing ",
    year = "2014",
    note = "Domain-Specific Languages and High-Level Frameworks for High-Performance Computing ",
    issn = "0743-7315",
-    doi = "https://doi.org/10.1016/j.jpdc.2014.07.003",
+    doi = {10.1016/j.jpdc.2014.07.003},
     url = "http://www.sciencedirect.com/science/article/pii/S0743731514001257",
  author = "H. Carter Edwards and Christian R. Trott and Daniel Sunderland"
 }
