### 2020-10-16 #2513 @eloisabentivegna
Hi @zingale, I have created a PR on the Castro repository with fixes to the DOI problems above. Can you check that everything works, and merge?
