### 2020-10-16 8af9f98
update the NSE test problem (#1231)
This revives the NSE test problem to use the new NSE network. We will use this for convergence testing with NSE + SDC
