+++ Exec/reacting_tests/nse_test/Prob_nd.F90
@@ -1,55 +1,36 @@
 subroutine amrex_probinit(init, name, namlen, problo, probhi) bind(c)
 
   use amrex_constants_module
-  use probdata_module, only: T_min, T_max, rho_ambient, width, xn, cfrac, ofrac
-  use network, only: network_species_index, nspec
-  use castro_error_module, only: castro_error
-  use amrex_fort_module, only: rt => amrex_real
+  use probdata_module
   use prob_params_module, only : center
+  use castro_error_module
+  use amrex_fort_module, only : rt => amrex_real
+  use eos_type_module, only : eos_t, eos_input_rt
+  use eos_module, only : eos
+  use extern_probin_module, only : small_x
+  use network
 
   implicit none
 
-  integer,  intent(in) :: init, namlen
-  integer,  intent(in) :: name(namlen)
+  integer, intent(in) :: init, namlen
+  integer, intent(in) :: name(namlen)
   real(rt), intent(in) :: problo(3), probhi(3)
 
-  integer, save :: ihe4, ic12, io16
 
-  real(rt) :: smallx
+  ! set explosion center
+  center(:) = ZERO
+  center(1) = HALF*(problo(1) + probhi(1))
+#if BL_SPACEDIM >= 2
+  center(2) = HALF*(problo(2) + probhi(2))
+#endif
+#if BL_SPACEDIM == 3
+  center(3) = HALF*(problo(3) + probhi(3))
+#endif
 
-  ! get the species indices
-  ihe4 = network_species_index("helium-4")
-  ic12 = network_species_index("carbon-12")
-  io16 = network_species_index("oxygen-16")
-
-  if (ihe4 < 0 .or. ic12 < 0 .or. io16 < 0) then
-     call castro_error("ERROR: species indices not found")
-  endif
-
-  ! make sure that the carbon fraction falls between 0 and 1
-  if (cfrac > 1.e0_rt .or. cfrac < 0.e0_rt) then
-     call castro_error("ERROR: cfrac must fall between 0 and 1")
-  endif
-
-  ! make sure that the oxygen fraction falls between 0 and 1
-  if (ofrac > 1.e0_rt .or. cfrac < 0.e0_rt) then
-     call castro_error("ERROR: ofrac must fall between 0 and 1")
-  endif
-
-  ! make sure that the C/O fraction sums to no more than 1
-  if (cfrac + ofrac > 1.e0_rt) then
-     call castro_error("ERROR: cfrac + ofrac cannot exceed 1.")
-  end if
-
-  ! set the default mass fractions
-  allocate(xn(nspec))
-
-  xn(:) = smallx
-  xn(ic12) = max(cfrac, smallx)
-  xn(io16) = max(ofrac, smallx)
-  xn(ihe4) = 1.e0_rt - cfrac - ofrac - (nspec - 2) * smallx
-
-  center(:) = HALF*(problo(:) + probhi(:))
+  ! we only work in NSE mode, so the mass fractions are not used by the EOS
+#ifndef NSE_THERMO
+  call castro_error("Error: this problem requires USE_NSE=TRUE")
+#endif
 
 end subroutine amrex_probinit
 
@@ -79,58 +60,84 @@ subroutine ca_initdata(level,time,lo,hi,nscal, &
                        state,state_lo,state_hi, &
                        delta,xlo,xhi)
 
-  use amrex_constants_module
-  use network, only: nspec
-  use eos_module, only: eos
-  use eos_type_module, only: eos_t, eos_input_rt
[…]
