+++ Exec/reacting_tests/nse_test/README.md
@@ -1,0 +1,8 @@
+# nse_test
+
+This is a simple test problem designed to explore how well hydro and
+reactions are coupled when a system enters NSE.
+
+This version is based on ``reacting_convergence`` (which is in turn
+based on ``acoustic_pulse_general``), but using the ``aprox19``
+network with the NSE table enabled.
