+++ paper/paper.bib
@@ -304,5 +304,6 @@ journal = "Journal of Parallel and Distributed Computing ",
       year = {2019},
     volume = {},
     number = {},
-     pages = {71-81}
+     pages = {71-81},
+       doi = {10.1109/P3HPC49587.2019.00012}
 }
