### 2020-10-20 #2513 @eloisabentivegna
The proof looks good to me! @zingale, I have added one last missing DOI, and issued a PR on the Castro repo. Once that is fixed, we should be done.
