### 2020-10-21 #2513 @eloisabentivegna
The proof looks fine now. I will issue the accept command next. Congratulations to the authors, and many heartfelt thanks to @kegiljarhus, @bonh, and @joshia5 for the careful reviews! 🎉
