### 2020-10-22 #2513 @kyleniemeyer
Hi @zingale, I'm the EIC on duty this week doing some final checks. Could you fix these in the paper before we accept?

- [ ] The paper is missing an explicit Statement of Need section, which we have made a [required component](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of articles. (You can likely use text already in the first page)
- [ ] In the second paragraph, can you use the `@author:2001` citation for the Collela 1990 reference, so that it appears as "Collela (1990)"
- [ ] please put a comma after "e.g." in paragraph four (and in any other places where i.e. or e.g. are used)
- [ ] The references are missing the journal names—likely this is due to the .bib file using astro-specific macros (`\apj`, etc.) that need to be replaced with spelled-out versions
