### 2020-10-23 9e5a203
add another resolution + a script to make a nice table
