### 2020-10-23 8a0c2de
add another resolution + a script to make a nice table (#1268)
